### 2023-10-23 #5978 @editorialbot
[PRE REVIEW]: Efficient Polyhedral Gravity Modeling in Modern C++ and Python
**Submitting author:** @schuhmaj (<a href="http://orcid.org/0009-0005-9693-4530">Jonas Schuhmacher</a>)
**Repository:** https://github.com/esa/polyhedral-gravity-model
**Branch with paper.md** (empty if default branch): main
**Version:** v2.1.0
