### 2023-10-24 d0d8588
add DOIs & URLs
