+++ paper/paper.bib
@@ -1,4 +1,5 @@
 @article{tsoulis2009recursive,
+  doi={10.1007/s00190-009-0310-9},
   title={Recursive algorithms for the computation of the potential harmonic coefficients of a constant density polyhedron},
   author={Tsoulis, Dimitrios and Jamet, Olivier and Verdun, J{\'e}r{\^o}me and Gonindard, Nicolas},
   journal={Journal of Geodesy},
@@ -9,6 +10,7 @@
 }
 
 @article{tsoulis2012analytical,
+  doi={10.1190/geo2010-0334.1},
   title={Analytical computation of the full gravity tensor of a homogeneous arbitrarily shaped polyhedral source using line integrals},
   author={Tsoulis, Dimitrios},
   journal={Geophysics},
@@ -20,6 +22,7 @@
 }
 
 @article{tsoulis2021computational,
+  doi={10.1111/1365-2478.13134},
   title={A computational review of the line integral analytical formulation of the polyhedral gravity signal},
   author={Tsoulis, Dimitrios and Gavriilidou, Georgia},
   journal={Geophysical Prospecting},
@@ -31,6 +34,7 @@
 }
 
 @article{izzo2022geodesy,
+  doi={10.1038/s44172-022-00050-3},
   title={Geodesy of irregular small bodies via neural density fields},
   author={Izzo, Dario and G{\'o}mez, Pablo},
   journal={Communications Engineering},
@@ -46,13 +50,15 @@
   author={Jonas Schuhmacher and Fabio Gratl and Dario Izzo and Pablo G{\'o}mez},
   booktitle={Proceedings of the 12th International Conference on Guidance, Navigation \& Control Systems (GNC)},
   year={2023},
+  url={https://az659834.vo.msecnd.net/eventsairwesteuprod/production-atpi-public/b1e566f6284f4814b1c733ec08e4d136}
 }
 
 @inproceedings{marak2023trajectory,
 	title={{Trajectory optimization of a spacecraft swarm orbiting around 67P/Churyumov-Gerasimenko}},
 	author={Mar{å}k, Rasmus and Blazquez, E. and G{\'o}mez, Pablo},
 	booktitle={Proceedings of the 9th International Conference on Astrodynamics Tools and Techniques, ICATT},
-	year={2023}
+	year={2023},
+  url={https://az659834.vo.msecnd.net/eventsairwesteuprod/production-atpi-public/6f457b8c46dd40ab826a9160c3110b56}
 }
 
 
@@ -61,9 +67,11 @@
 	author={Schuhmacher, Jonas},
 	year={2022},
 	institution={Technische Universität München},
+  url={https://mediatum.ub.tum.de/doc/1695208/1695208.pdf}
 }
 
 @article{tsoulis2001singularities,
+  doi={10.1190/1.1444944},
   title={On the singularities of the gravity field of a homogeneous polyhedral body},
   author={Tsoulis, Dimitrios and Petrovi{\'c}, Sveto},
   journal={Geophysics},
@@ -75,6 +83,7 @@
 }
 
 @article{petrovic1996determination,
+  doi={10.1007/s001900050074},
   title={Determination of the potential of homogeneous polyhedral bodies using line integrals},
   author={Petrovi{\'c}, S},
   journal={Journal of Geodesy},
@@ -85,6 +94,7 @@
 }
 
 @article{hang2015tetgen,
+  doi={10.1145/2629697},
   title={TetGen, a Delaunay-based quality tetrahedral mesh generator},
   author={Hang, Si},
   journal={ACM Trans. Math. Softw},
@@ -104,6 +114,7 @@
 }
 
 @article{vsprlak2021use,
+  doi={10.1016/j.earscirev.2021.103739},
[…]
