### 2023-10-24 5707d40
eros is now eros and not 67p
