### 2023-10-25 #5978 @schuhmaj
> @schuhmaj — Thanks for your submission! I am working to assign an editor, but in the meantime can you please check those DOIs recommended by the bot above ☝️ and add them to your bibliography if appropriate.

@dfm Thank you very much. I've reviewed and added the missing DOIs 👍
