### 2024-02-04 #5978 @dfm
👋 @santisoler, @mikegrudic — Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
