### 2024-02-20 #6384 @editorialbot
[REVIEW]: Efficient Polyhedral Gravity Modeling in Modern C++ and Python
**Submitting author:** @schuhmaj (<a href="http://orcid.org/0009-0005-9693-4530">Jonas Schuhmacher</a>)
**Repository:** https://github.com/esa/polyhedral-gravity-model
**Branch with paper.md** (empty if default branch): main
**Version:** 3.2.0
