### 2024-03-06 esa/polyhedral-gravity-model#19 @santisoler
[JOSS Review] Link to FORTRAN implementation is not working
While reading the `README.md` I noticed that the link to Tsoulis' FORTRAN implementation is not working. At least I cannot access that code. Is this the right url? Or this is a problem on the SEG? I'm referring to this line:

https://github.com/esa/polyhedral-gravity-model/blob/ebe47ecb30b946eb9a260702a096fb849611cf4d/README.md?plain=1#L55

This is not a major concern, but I decided to open this Issue it so you are aware.

---

This issue is part of the JOSS review: openjournals/joss-reviews#6384
