### 2024-03-07 #6384 @santisoler
Hello @schuhmaj. Thanks for submitting this paper to JOSS. As you might have noticed I've started with the review process. I still need to go through some of the items on the checklist. While I do so I'll be opening issues and PRs in your repo (as the ones you can already see). Feel free to reply and tackle them in whatever order you like. They won't block my review process: even if I open an issue, I'll continue with the rest of the checklist.

The suggestions I make for solving the issues I identify and the PRs I open are mere suggestions to solve them. Don't feel obliged to accept them if you don't agree with my take. Feel free to suggest alternative ways of solving them, and to justify  your disagreement if you have any.

## On contribution and authorship

I do have a question that I'll ask you in this Issue. I noticed that `gomezzz` made significant contributions to the submitted repository, particularly in https://github.com/esa/polyhedral-gravity-model/commit/f3cc4d9d80c3ca23fffa342d9ac178a4582c0899. Is there a reason why `gomezzz` is not listed in the list of submitting authors?

I'm not asking them to be added, I'm just curious about their involvement in the submitted paper.

(I'm not pinging `gomezzz` to avoid sending unrequested notifications. I leave the decision to ping them or not to the editor @dfm).
