### 2024-03-07 #6384 @schuhmaj
Hi @santisoler,  Thanks a lot for taking the time to do our JOSS review. I will address the issues that have already been opened and those that will come up in the forthcoming days.

Regarding the authorship, the project was initially developed in [schuhmaj/polyhedral-gravity-model](https://github.com/schuhmaj/polyhedral-gravity-model)
The [`init`](https://github.com/esa/polyhedral-gravity-model/commit/f3cc4d9d80c3ca23fffa342d9ac178a4582c0899) commit from `gomezzz` roughly corresponds to the final commit in [schuhmaj/polyhedral-gravity-model](https://github.com/schuhmaj/polyhedral-gravity-model).
We did not transfer the repository at that time to keep a copy of the "original," while the ESA repository started with a fresh commit history.

`gomezzz` has been (and sort of still is 🙂 since he suggested the JOSS submission) my major advisor from the ESA side through code reviews and suggestions.
