### 2024-03-07 #6384 @gomezzz
Hi @santisoler , thanks for reviewing ! To chime in from my end and just confirm, indeed, the majority of the code has actually been written by @schuhmaj . My role has been more on the side of ESA interface and code review as you can see in pull requests in the repo pointed out by @schuhmaj . :)
