### 2024-03-07 #6384 @santisoler
Thanks @schuhmaj for the clarification, and thanks @gomezzz for confirming your role in the submitted repository. All good now on the **Contribution and authorship** end, I've already checked off that box in my checklist.

I'll continue with my review in the following days. Looking forward to it.
