### 2024-03-07 esa/polyhedral-gravity-model#22 @santisoler
[JOSS Review] Missing Ninja requirement for building from sources
## The issue

The instructions to build `polyhedral-gravity-model` from sources mention CMake as a requirement. Nevertheless, [Ninja](https://ninja-build.org/) also is required, otherwise the build process runs into an error:

```
$ pip install .
Processing /home/santi/git/polyhedral-gravity-model
  Preparing metadata (setup.py) ... done
Building wheels for collected packages: polyhedral_gravity
  Building wheel for polyhedral_gravity (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [56 lines of output]
      running bdist_wheel
      running build
      running build_ext
      CMake Error: CMake was unable to find a build program corresponding to "Ninja".  CMAKE_MAKE_PROGRAM is not set.  You probably need to select a different build tool.
      CMake Error: CMAKE_C_COMPILER not set, after EnableLanguage
      CMake Error: CMAKE_CXX_COMPILER not set, after EnableLanguage
      -- Configuring incomplete, errors occurred!
      See also "/home/santi/git/polyhedral-gravity-model/build/temp.linux-x86_64-cpython-311/polyhedral_gravity/CMakeFiles/CMakeOutput.log".
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "/home/santi/git/polyhedral-gravity-model/setup.py", line 148, in <module>
          setup(
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/__init__.py", line 103, in setup
          return distutils.core.setup(**attrs)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/core.py", line 185, in setup
          return run_commands(dist)
                 ^^^^^^^^^^^^^^^^^^
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/core.py", line 201, in run_commands
          dist.run_commands()
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 969, in run_commands
          self.run_command(cmd)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/dist.py", line 963, in run_command
          super().run_command(command)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 988, in run_command
          cmd_obj.run()
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/wheel/bdist_wheel.py", line 368, in run
          self.run_command("build")
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/cmd.py", line 318, in run_command
          self.distribution.run_command(command)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/dist.py", line 963, in run_command
          super().run_command(command)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 988, in run_command
          cmd_obj.run()
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/command/build.py", line 131, in run
          self.run_command(cmd_name)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/cmd.py", line 318, in run_command
          self.distribution.run_command(command)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/dist.py", line 963, in run_command
          super().run_command(command)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 988, in run_command
          cmd_obj.run()
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/command/build_ext.py", line 89, in run
          _build_ext.run(self)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/command/build_ext.py", line 345, in run
          self.build_extensions()
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/command/build_ext.py", line 467, in build_extensions
          self._build_extensions_serial()
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/site-packages/setuptools/_distutils/command/build_ext.py", line 493, in _build_extensions_serial
          self.build_extension(ext)
        File "/home/santi/git/polyhedral-gravity-model/setup.py", line 138, in build_extension
          subprocess.check_call(["cmake", ext.sourcedir] + cmake_args, cwd=build_temp)
        File "/home/santi/.mambaforge/envs/polyhedral-gravity-env/lib/python3.11/subprocess.py", line 413, in check_call
          raise CalledProcessError(retcode, cmd)
      subprocess.CalledProcessError: Command '['cmake', '/home/santi/git/polyhedral-gravity-model', '-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=/home/santi/git/polyhedral-gravity-model/build/lib.linux-x86_64-cpython-311/', '-DPYTHON_EXECUTABLE=/home/santi/.mambaforge/envs/polyhedral-gravity-env/bin/python3.11', '-DCMAKE_BUILD_TYPE=Release', '-DPOLYHEDRAL_GRAVITY_PARALLELIZATION=TBB', '-DLOGGING_LEVEL=2', '-DUSE_LOCAL_TBB=OFF', '-DBUILD_POLYHEDRAL_GRAVITY_DOCS=OFF', '-DBUILD_POLYHEDRAL_GRAVITY_TESTS=OFF', '-DBUILD_POLYHEDRAL_GRAVITY_LIBRARY=OFF', '-DBUILD_POLYHEDRAL_GRAVITY_EXECUTABLE=OFF', '-DBUILD_POLYHEDRAL_GRAVITY_PYTHON_INTERFACE=ON', '-DBUILD_SHARED_LIBS=OFF', '-GNinja', '--log-level=VERBOSE']' returned non-zero exit status 1.
      [end of output]

  ERROR: Failed building wheel for polyhedral_gravity
  Running setup.py clean for polyhedral_gravity
Failed to build polyhedral_gravity
ERROR: Could not build wheels for polyhedral_gravity, which is required to install pyproject.toml-based projects
```

## The solution

I would recommend to add Ninja to the list of requirements needed to build the package from sources. The change should be applied to the `README.md` and the `docs/build.rst` files.

---

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/6384
