### 2024-03-07 02c0a74
Minor fix to example using GravityEvaluable
Replace `model.GravityEvaluable` for
`polyhedral_gravity.GravityEvaluable` in the minimal Python example
showcased in the `README.md`.
