### 2024-03-07 esa/polyhedral-gravity-model#24 @santisoler
[JOSS Review] Improve discoverability of the documentation pages
In the `README.md` we can find a [link to the documentation pages](https://polyhedral-gravity-model-cpp.readthedocs.io/en/stable/) hosted in ReadTheDocs. I think these pages are super helpful when getting familiar with the package.
Nonetheless, this link is not easily discoverable.

Would be nice of a link to it were more discoverable. For example, by adding it to the Website field in the **About** section of the repo (checkout the screen captures in https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/classifying-your-repository-with-topics#adding-topics-to-your-repository for reference):

![image](https://github.com/esa/polyhedral-gravity-model/assets/11541317/ebbf0a14-0f75-470e-826d-fa50e220e85c)

![bitmap](https://github.com/esa/polyhedral-gravity-model/assets/11541317/24c2142b-1403-4aa0-bf27-c56aa7babac1)

Also, the link could be added to the top section of the `README.md`, in a similar way as [Numpy does](https://github.com/numpy/numpy).

---

This issue is part of the JOSS review: openjournals/joss-reviews#6384
