### 2024-03-07 esa/polyhedral-gravity-model#21 @santisoler
I just noticed by comparing the example with the [Jupyter Notebook](https://github.com/esa/polyhedral-gravity-model/blob/66bc580135b8a224df4b2c5b4a0fd4ab3cdd2e10/script/polyhedral-gravity.ipynb) that if we define the `computation_point` as a 1d array (`computation_point = np.array([0, 0, 0])`) the example works fine.

This version of it doesn't fail on my end:

```python
import numpy as np
import polyhedral_gravity

# We define the cube as a polyhedron with 8 vertices and 12 triangular faces
# The density is set to 1.0
cube_vertices = np.array(
    [
        [-1, -1, -1],
        [1, -1, -1],
        [1, 1, -1],
        [-1, 1, -1],
        [-1, -1, 1],
        [1, -1, 1],
        [1, 1, 1],
        [-1, 1, 1],
    ]
)
cube_faces = np.array(
    [
        [1, 3, 2],
        [0, 3, 1],
        [0, 1, 5],
        [0, 5, 4],
        [0, 7, 3],
        [0, 4, 7],
        [1, 2, 6],
        [1, 6, 5],
        [2, 3, 6],
        [3, 7, 6],
        [4, 5, 6],
        [4, 6, 7],
    ]
)
cube_density = 1.0
computation_point = np.array([0, 0, 0])

potential, acceleration, tensor = polyhedral_gravity.evaluate(
    polyhedral_source=(cube_vertices, cube_faces),
    density=cube_density,
    computation_points=computation_point,
    parallel=True,
)
```
