### 2024-03-07 esa/polyhedral-gravity-model#22 @santisoler
Minor comment regarding this issue. Ninja is needed to build the sources and install the Python interface and building the package, as instructed here: https://polyhedral-gravity-model-cpp.readthedocs.io/en/stable/build.html#build-installation-from-source

Building the C++ binaries from source (following [these instructions](https://polyhedral-gravity-model-cpp.readthedocs.io/en/stable/build.html#build-with-cmake)) works fine without Ninja installed.
