### 2024-03-07 3e8998a
Fix Python tests to work with latest Numpy version
Modify how the potential and acceleration are retrieved from the
solution returned by `evaluate()` and `GravityEvaluable` in tests so
the code can run with latest Numpy versions. The previous tests were
trying to generate an array from ragged tuples, which is no longer
supported by Numpy. Also, modify the `assert
np.testing.assert_almost_equal() is None` lines since the Numpy testing
functions always return `None`.
