### 2024-03-08 esa/polyhedral-gravity-model#27 @santisoler
[JOSS Review] Improve tests by using densities different than 1.0
## The issue

While reading the C++ and the Python tests, I noticed that the ones that compare the results for a cube against the analytic solutions define the density of the cube to be equal to 1.0:

https://github.com/esa/polyhedral-gravity-model/blob/ebe47ecb30b946eb9a260702a096fb849611cf4d/test/python/test_polyhedral_gravity.py#L34

https://github.com/esa/polyhedral-gravity-model/blob/ebe47ecb30b946eb9a260702a096fb849611cf4d/test/calculation/GravityModelCubeTest.cpp#L29

Choosing a density equals to 1.0 that the test functions aren't checking if the density value is properly used in the code.

In the hypothetical case that a bug is introduced into this line:

https://github.com/esa/polyhedral-gravity-model/blob/ebe47ecb30b946eb9a260702a096fb849611cf4d/src/polyhedralGravity/calculation/GravityEvaluable.cpp#L65

to something like:

```cpp
        const double prefix = util::GRAVITATIONAL_CONSTANT / _density;
```

then all tests will still pass, not being able to flag this bug.

## The solution

I suggest to set the density of these cubes to a different value that could flag such bug.

---

This issue is part of the JOSS review: openjournals/joss-reviews#6384
