### 2024-03-08 esa/polyhedral-gravity-model#28 @santisoler
[JOSS Review] Sign of acceleration components?
While playing with the minimal Python example shown in the `README.md` I defined a cube in the (+,+,+) octant, so one of its vertices is in `[1, 1, 1]` and its opposite one in `[2, 2 ,2]`, with a density of 1.0. Then I defined a computation point in the origin (`[0, 0, 0]`) and used the `evaluate()` function to compute its fields.

```python
import numpy as np
import polyhedral_gravity

cube_vertices = np.array(
    [
        [1, 1, 1],
        [2, 1, 1],
        [2, 2, 1],
        [1, 2, 1],
        [1, 1, 2],
        [2, 1, 2],
        [2, 2, 2],
        [1, 2, 2],
    ]
)
cube_faces = np.array(
    [
        [1, 3, 2],
        [0, 3, 1],
        [0, 1, 5],
        [0, 5, 4],
        [0, 7, 3],
        [0, 4, 7],
        [1, 2, 6],
        [1, 6, 5],
        [2, 3, 6],
        [3, 7, 6],
        [4, 5, 6],
        [4, 6, 7],
    ]
)
cube_density = 1.0
computation_point = np.array([0, 0, 0])

potential, acceleration, tensor = polyhedral_gravity.evaluate(
    polyhedral_source=(cube_vertices, cube_faces),
    density=cube_density,
    computation_points=computation_point,
    parallel=True,
)
print(acceleration)
```

After running this code I get:
```
[-5.715199089414289e-12, -5.715199089414289e-12, -5.7151990894142846e-12]
```

The three components of the acceleration are negative, which would imply that any particle located in the computation point would feel an acceleration towards the (-,-,-) octant, meaning it would feel a repulsive force from the gravitational interaction with the cube. Given that the cube has a positive density, I'd assume that the three components should be positive, not negative.

Is this the expected result? Or are there any sign conventions that I missed?

---

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/6384
