### 2024-03-08 esa/polyhedral-gravity-model#28 @JoachimSchwabe
The output is the derivative of the potential, not the gravitational acceleration. Since the potential becomes smaller at further distance of the body, the sign of dV/d(x,y,z) is negative. The gravitational acceleration has the same value but with flipped sign, e.g. dg_x = -dV/dx.
