### 2024-03-10 esa/polyhedral-gravity-model#30 @mikegrudic
[JOSS Review] Documentation missing statement of need.
Just a small box to tick: for JOSS, the documentation must have:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I don't see such a thing in the current [documentation](https://polyhedral-gravity-model-cpp.readthedocs.io/en/stable/build.html) - just a brief paragraph about what the code does and is intended for would be fine.

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/6384
