### 2024-03-10 #6384 @mikegrudic
Hi @schuhmaj,

Just did a pass through the code. Overall I think the software and paper are of high quality. I raised one issue re: documentation, but that's the one thing I would need to see to sign off on this. However, I strongly concur with all of @santisoler's issues and PRs and recommend addressing them.

Cheers,
Mike
