### 2024-03-11 esa/polyhedral-gravity-model#30 @santisoler
I agree with this Issue. I also think some statement of what `polyhedral-gravity-model` is, what problems does it solve, and how is the intended audience should be added to the documentation (and also to the `README.md` if that's desirable).
