### 2024-03-13 7f1e4fa
apply README.md fixes to docs
