### 2024-03-13 esa/polyhedral-gravity-model#19 @schuhmaj
For me, as well, the link does not work anymore. Thank you for letting me know.

I will leave this issue open for now. Maybe the site will return in a few days/ weeks. Otherwise, I remove the broken link.
