### 2024-03-13 esa/polyhedral-gravity-model#19 @santisoler
Sounds good. FIY, this issue will not block the publication of the JOSS article. I just wanted to let you know that the link is broken.
