+++ docs/quick_start_python.rst
@@ -52,14 +52,15 @@ defined from within source code for a specific point and density.
         import polyhedral_gravity as model
 
         # Defining every input parameter in the source code
-        vertices = ...          # [] of [] or np.array of length 3 and type float
-        faces = ...             # [] of [] or np.array of length 3 and type int
+        vertices = ...          # (N-3)-array-like of type float
+        faces = ...             # (N-3)-array-like of type int
         density = ...           # float
-        computation_point = ... # [] or np.array of length 3
+        computation_point = ... # (3)-array-like
 
         # Evaluate the gravity model
         # Notice that the third argument could also be a list of points
         # Returns a tuple of potential, acceleration and tensor
+        # If computation_point would be a (N,3)-array, the output would be list of triplets!
         potential, acceleration, tensor = model.evaluate((vertices, faces), density, computation_point, parallel=True)
 
 
@@ -75,7 +76,7 @@ Of course, you can also use keyword arguments for the parameters.
         file_vertices = '___.node'      # str, path to file
         file_nodes = '___.face'         # str, path to file
         density = ...                   # float
-        computation_points = ...        # [] or np.array of length 3
+        computation_points = ...        # (N,3)-array-like
 
         # Evaluate the gravity model
         # Notice that the last argument could also be a list of points
@@ -98,11 +99,12 @@ in some source files for a specific point and density.
         # Reading the vertices and files from a single .mesh file
         file = '___.mesh'       # str, path to file
         density = ...           # float
-        computation_point = ... # [] or np.array of length 3
+        computation_point = ... # (3)-array-like
 
         # Evaluate the gravity model
         # Notice that the last argument could also be a list of points
         # Returns a tuple of potential, acceleration and tensor
+        # If computation_point would be a (N,3)-array, the output would be list of triplets!
         potential, acceleration, tensor = model.evaluate([mesh], density, computation_point)
 
 
@@ -120,17 +122,19 @@ about the vertices' ordering due to its quadratic complexity!
     import polyhedral_gravity.utility as mesh_sanity
 
     # Defining every input parameter in the source code
-    vertices = ...              # [] of [] or np.array of length 3 and type float
-    faces = ...                 # [] of [] or np.array of length 3 and type int
-    density = ...               # float
-    computation_point = ...     # [] or np.array of length 3
+    vertices = ...          # (N-3)-array-like of type float
+    faces = ...             # (N-3)-array-like of type int
+    density = ...           # float
+    computation_point = ... # (3)-array-like
 
-    # Evaluate the gravity model
-    # Notice that the last argument could also be a list of points
-    # Returns (list of) tuple of potential, acceleration and tensor
 
-    # Additional guard statement
+
+    # Additional guard statement to check that the plane normals
+    # are outwards pointing
     if mesh_sanity.check_mesh(vertices, faces):
+        # Evaluate the gravity model
+        # Returns a tuple of potential, acceleration and tensor
+        # If computation_point would be a (N,3)-array, the output would be list of triplets!
         potential, acceleration, tensor = model.evaluate((vertices, faces), density, computation_point)
 
 
@@ -150,17 +154,22 @@ Have a look at the example below to see how to use the :code:`GravityEvaluable` 
         import polyhedral_gravity as model
 
         # Defining every input parameter in the source code
-        vertices = ...           # [] of [] or np.array of length 3 and type float
-        faces = ...              # [] of [] or np.array of length 3 and type int
+        vertices = ...           # (N-3)-array-like of type float
+        faces = ...              # (N-3)-array-like of type int
         density = ...            # float
-        computation_points = ... # [] or np.array of length 3
+        computation_points = ... # (N,3)-array-like
 
         # Create the evaluable object
[…]
