### 2024-03-18 4ac841d
resolve unnecessary ninja dependency
