### 2024-03-20 esa/polyhedral-gravity-model#28 @schuhmaj
## The problem

So, the primary concern is that the acceleration's sign is inverted compared to the potential and the second derivative tensor.
Hence, the potential fix is the simple multiplication with -1.
I totally agree with that, especially given that, as @JoachimSchwabe  reported, the numerical difference of the potential yields a value with an inverted sign to the returned acceleration.

## Handling until now

However, I want to put the following things for reference in the issue (since this sign question still is a bit unclear to me since I am not a domain expert):
- Tsoulis et al.'s reference implementation does not mention how to handle the sign. He returns the [accelerations in his Fortran and Matlab Implementations as absolute values](https://github.com/Gavriilidou/GPolyhedron/blob/81060688383a36bceccb9d2a9b89bf20f6532763/GPolyhedron.m#L734).
- The [analytical cube gravity](https://arxiv.org/pdf/1206.3857.pdf), which we used as a reference, defines the derivative of the cube as the negative derivative of the potential. However, the potential is defined as negative in the first place. (So, this is aligned to our acceleration but inverted to our potential - One of my advisors meant at that time that this is just a "definition thing.")
- We utilized this polyhedral gravity model as the ground truth for neural density fields. Here, potential and acceleration were [both multiplied with -1](https://github.com/gomezzz/geodesyNets/blob/5122889d315eacb1f2a6ecd81b753a2205c0cda7/gravann/labels/polyhedral.py#L55). This yielded the same results as the Mascon Model utilized in a previous study iteration.

## Towards a solution

To summarize, I am highly convinced by the fact that the numerical derivatives sign is flipped when compared to the derivative as returned by the implementation.

So, **I would definitely add a minus multiplication to the acceleration output.**
@santisoler @JoachimSchwabe  **I would still ask you for your confirmation that this is the change you desire?**

----
Anyways, many thanks the two of you for bringing up this issue in detail and stating examples - this is highly appreciated.
