### 2024-03-20 esa/polyhedral-gravity-model#32 @santisoler
[JOSS Review] Minor improvement in paper paragraph
## The issue

While reading the submitted paper to JOSS, I noticed that the first paragraph of the "Statement of Need" section could be improved to make it more clear:

https://github.com/esa/polyhedral-gravity-model/blob/fa2dba34709a42dec29e3fb3227c83f161693bc4/paper/paper.md?plain=1#L48

> The complex gravitational fields of irregular bodies, such as asteroids and comets, are often modeled using polyhedral gravity models since alternative approaches like mascon models or spherical harmonics struggle with these bodies' irregular geometry. The former struggles with convergence close to the surface [@vsprlak2021use], whereas the latter requires a computationally expensive amount of mascons (point masses of which the target body is composed) to model fine-granular surface geometry [@wittick2017mascon].

I think the usage of _former_ and _latter_ is not clear to which methods they are referring to. From the context, I understand that the _former_ should refer to the spherical harmonics approach, while the _latter_ should refer to the mascon models. If that's the case, the order of those are interchanged. Nonetheless, using _former_ here might lead the reader to think it's referring to the _polyhedral gravity models_ and not to one of the other methods.

## Possible solution

I would suggest to modify that paragraph, explicitly mentioning the methods the sentence is referring to. Something along:

```markdown
The complex gravitational fields of irregular bodies, such as asteroids and
comets, are often modeled using polyhedral gravity models since alternative
approaches like mascon models or spherical harmonics struggle with these
bodies' irregular geometry. The spherical harmonics approach struggles with
convergence close to the surface [@vsprlak2021use], whereas the mascon models
require a computationally expensive amount of mascons (point masses of which
the target body is composed) to model fine-granular surface geometry
[@wittick2017mascon].
```

I think it's ok to repeat the names of the methods in favor of clarity.

Please take this as a suggestion, and feel free to modify the paragraph at will.

---

This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/6384
