### 2024-03-20 esa/polyhedral-gravity-model#33 @santisoler
[JOSS Review] Provide a list of dependencies needed to run the Jupyter Notebook
## The Issue

While trying to run the provided Jupyter Notebook locally, I noticed several dependencies need to be installed in the environment so it can actually run properly. Here I copy a non-comprehensive list of packages that are needed to run the notebook:

- `jupyter`
- `matplotlib`
- `sympy`
- `pyvista`
- `meshio`
- `tetgen`
- `openmesh`

## Possible solution

Would be nice to include a list of required dependencies in the `README.md`, or through an `environment.yml`.

---

This issue is part of the JOSS review: openjournals/joss-reviews#6384
