### 2024-03-20 #6384 @santisoler
Hi @schuhmaj. Thanks again for submitting this article to JOSS. I finished my review process. I think this article is in a very good shape to be published. I would only ask the authors to address some of the issues mentioned in the opened Issues and Pull Requests, which have been linked to this Review Issue.

Regarding the checklist, I think the authors have covered all points, except for the one mentioned by @mikegrudic related to a missing statement of need in the documentation (https://github.com/esa/polyhedral-gravity-model/issues/30).

In conclusion, I think the article should be considered as ready for publication after these issues have been solved. I really enjoyed reviewing this article and going through the code, and I'm looking forward to continuing the conversation.

Kind regards,
Santiago Soler
