### 2024-03-20 esa/polyhedral-gravity-model#28 @santisoler
Hi @schuhmaj. Thanks for the detailed reply. I think the solution of adding a -1 to the acceleration will solve the issue I get in that particular example, although I think it would be more important to be sure why the implementation needs that -1 sign, so the fix applies to every possible case and not just this one.

Regarding your comments on why this is needed, I'm not very familiar with Tsoulis et al.'s implementation, and I'm not sure why they return absolute values. But I could say that your advisor is right when they stated that the issue the sign of the potential is a "definition thing".

Lets consider the gravity acceleration vector $\mathbf{g}(\mathbf{p})$ on an observation point $\mathbf{p}$ due to a particle located on $\mathbf{q}$ with mass $m$. If $m>0$, then this vector will always be attractive, therefore pointing in the $\mathbf{q} - \mathbf{p}$ direction. Since gravitational forces are [conservative](https://en.wikipedia.org/wiki/Conservative_force), we can define a potential $V$ such as:

$$ \mathbf{g} = - \nabla V $$

The minus sign used in the definition of $V$ is a convention. Mostly in Physics, potentials are defined with the minus sign, so negative values of the potential mean attractive fields, while positive values mean repulsive fields. This is why in most physical sciences (including the [analytical cube gravity](https://arxiv.org/pdf/1206.3857.pdf) article you shared), the gravitational potential is always negative for $m>0$.

Nonetheless, in geodesy and geophysics, the positive sign convention is usually used, so $V$ is defined as (Hofmann-Wellenhof and Moritz, 2005, eq. 1-9):

$$ \mathbf{g} = + \nabla V $$

Meaning that gravitational potentials have positive values for $m>0$ bodies. This convention doesn't impact the sign of the acceleration vector components (which are physical observable quantities that don't depend on sign conventions) nor on the signs of the tensor components.

In conclusion, the sign of the potential could be either positive or negative, as long as it is clear what convention the implementation is applying. Regarding the acceleration vector, its sign must match the attractive nature of gravitational interactions for positive masses.

---

Hofmann-Wellenhof and Moritz (2005). Physical Geodesy. doi: [10.1007/b139113](https://doi.org/10.1007/b139113)
