### 2024-03-22 esa/polyhedral-gravity-model#28 @JoachimSchwabe
I think I have to make a general comment here:

Firstly, as I now realized, with the correct mesh orientation of the test cube the package delivers a positive potential (outside the test body), decreasing when facing away from the mass. As outlined below by @santisoler, this is in line with the convention in physical geodesy.

Now, regarding this geodetic convention:
1. Indeed in physical geodesy, the potential of a mass with positive density is always considered positive. It is understood as the work to be done against the gravitational force to bring a test mass away from the body to infinity.
2. Yes, the actual physical (measurable) acceleration (i.e., length of the potential gradient vector | grad V |) is positive and facing towards the mass.
3. However, for the partial derivatives Tx/Ty/Tz of the potential ("acceleration"), we have to consider the location of the computation point with respect to the mass in the coordinate system. In gravimetry, the observation/computation point is usually above the mass (z_P > 0 and z_mass < z_P, z is upwards). This means that Tz of the positive potential is negative in the computation point (V decreasing upwards). On the other hand, what is measured with gravimeters is the actual acceleration acting downwards. This is why in gravimetry, per definition acc_z = -Tz! The tensor element Tzz = d/dz(Tz) = -d/dz(acc_z) is then again positive, since the negative Tz becomes smaller upwards (away from the mass). If we now assume the test cube hovering in the air and the computation point is below the cube (z_P < 0, z_mass > z_P), consequently Tz must become positive along the z axis (towards the cube). But because of the convention acc_z = -Tz, the code will output a negative acc_z in this case. Tzz is again positive, because the positive Tz becomes larger along the z axis (towards the mass).
4. The same also holds for all the other coordinate axis. The "acceleration" is defined as the negative of Tx/Ty/Tz, which both can be positive or negative depending whether the respective axis is facing towards or away from the mass! This makes sense because if we consider any body in a free xyz coordinate system, in fact there is no "up", "left" or "right". Only that the implementation of the package (right-handed coordinate system) happens to match with geodetic coordinates such as e.g. UTM coordinates: "East" (x), "North" (y), Height (z).
