### 2024-03-22 esa/polyhedral-gravity-model#34 @santisoler
Check validity of the polyhedra before computing gravity fields?
While reading some of the in #29, I noticed that the `evaluate()` function is not checking if the passed polyhedra is valid or not, and returning the gravity field values without any warning that they might not be right.

For example, grabbing the cube I used as example in #28, the following code runs with a valid cube and returns these values:

```python
import numpy as np
import polyhedral_gravity

cube_vertices = np.array(
    [
        [1, 1, 1],
        [2, 1, 1],
        [2, 2, 1],
        [1, 2, 1],
        [1, 1, 2],
        [2, 1, 2],
        [2, 2, 2],
        [1, 2, 2],
    ]
)
cube_faces = np.array(
    [
        [1, 3, 2],
        [0, 3, 1],
        [0, 1, 5],
        [0, 5, 4],
        [0, 7, 3],
        [0, 4, 7],
        [1, 2, 6],
        [1, 6, 5],
        [2, 3, 6],
        [3, 7, 6],
        [4, 5, 6],
        [4, 6, 7],
    ]
)
cube_density = 1.0
computation_point = np.array([0, 0, 0])

print("Valid cube?", polyhedral_gravity.utility.check_mesh(cube_vertices, cube_faces))

potential, acceleration, tensor = polyhedral_gravity.evaluate(
    polyhedral_source=(cube_vertices, cube_faces),
    density=cube_density,
    computation_points=computation_point,
    parallel=True,
)

print(f"potential: {potential}")
print(f"acceleration: {acceleration}")
print(f"tensor: {tensor}")
```

```
Valid cube? True
potential: 2.5695100395128033e-11
acceleration: [-5.715199089414289e-12, -5.715199089414289e-12, -5.7151990894142846e-12]
tensor: [2.2229884599767045e-26, 2.2229884599767045e-26, 2.2229884599767045e-26, 3.8189716252752374e-12, 3.818971625275311e-12, 3.8189716252752374e-12]
```

But, if I flip the order of the vertices indices in the first row of `cube_faces`, then the results look like these:

```python
import numpy as np
import polyhedral_gravity

cube_vertices = np.array(
    [
        [1, 1, 1],
        [2, 1, 1],
        [2, 2, 1],
        [1, 2, 1],
        [1, 1, 2],
        [2, 1, 2],
        [2, 2, 2],
        [1, 2, 2],
    ]
)
cube_faces = np.array(
    [
        [1, 2, 3],
        [0, 3, 1],
        [0, 1, 5],
        [0, 5, 4],
        [0, 7, 3],
        [0, 4, 7],
        [1, 2, 6],
        [1, 6, 5],
        [2, 3, 6],
        [3, 7, 6],
        [4, 5, 6],
        [4, 6, 7],
    ]
)
cube_density = 1.0
computation_point = np.array([0, 0, 0])

print("Valid cube?", polyhedral_gravity.utility.check_mesh(cube_vertices, cube_faces))

potential, acceleration, tensor = polyhedral_gravity.evaluate(
    polyhedral_source=(cube_vertices, cube_faces),
    density=cube_density,
    computation_points=computation_point,
    parallel=True,
)

print(f"potential: {potential}")
print(f"acceleration: {acceleration}")
print(f"tensor: {tensor}")
```

```
Valid cube? False
potential: 3.8680139904499555e-11
acceleration: [-5.715199089414289e-12, -5.715199089414289e-12, 2.025487992932876e-11]
tensor: [2.2229884599767045e-26, 2.2229884599767045e-26, -3.96702361734801e-12, 3.8189716252752374e-12, 3.818971625275311e-12, 3.8189716252752374e-12]
```

As you can see, the values of $g_z$ and $g_{zz}$ are significantly different between the two outputs.

I think it would be a good idea to include a validity check in the `evaluate()` function before computing the gravity fields. This would prevent users getting wrong results and therefore making wrong decisions or arriving at wrong conclusions because their results were wrong.

---

This issue is unrelated to the JOSS review. Addressing it won't be required for the JOSS paper to be published, but I think making the `evaluate()` function less error prone would be a major feature for this package.
