+++ .github/workflows/docs.yml
@@ -1,0 +1,18 @@
+name: Deploy Sphinx documentation to Pages
+
+on:
+  push:
+    branches: ["main"]
+
+jobs:
+  pages:
+    runs-on: ubuntu-latest
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    permissions:
+      pages: write
+      id-token: write
+    steps:
+      - id: deployment
+        uses: sphinx-notes/pages@v3
