+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 breathe
+sphinx-book-theme
