+++ README.md
@@ -1,7 +1,7 @@
 # polyhedral-gravity-model
 
-![Build and Test](https://github.com/esa/polyhedral-gravity-model/actions/workflows/build-test.yml/badge.svg)
-![Documentation Status](https://github.com/esa/polyhedral-gravity-model/actions/workflows/docs.yml/badge.svg)
+![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/esa/polyhedral-gravity-model/.github%2Fworkflows%2Fbuild-and-test.yml?label=Build%20and%20Test)
+![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/esa/polyhedral-gravity-model/.github%2Fworkflows%2Fdocs.yml?logo=GitBook&label=Documentation)
 ![GitHub](https://img.shields.io/github/license/esa/polyhedral-gravity-model)
 
 ![PyPI](https://img.shields.io/pypi/v/polyhedral-gravity)
