+++ .github/workflows/docs.yml
@@ -4,6 +4,9 @@ on:
   push:
     branches: ["main"]
 
+env:
+  GITHUB_PAGES_BUILD: True
+
 jobs:
   pages:
     runs-on: ubuntu-latest
