+++ docs/conf.py
@@ -30,26 +30,12 @@ def configure_doxyfile(input_dir, output_dir):
         file.write(filedata)
 
 
-# Check if we're running on Read the Docs' servers
-read_the_docs_build = os.environ.get("READTHEDOCS", None) == "True"
-
 breathe_projects = {}
-if read_the_docs_build:
-    input_dir = "../src"
-    output_dir = "build"
-    configure_doxyfile(input_dir, output_dir)
-    subprocess.call("doxygen", shell=True)
-    breathe_projects["polyhedral-gravity-model"] = output_dir + "/xml"
-
-
-github_pages_build = os.environ.get("GITHUB_PAGES_BUILD", None) == "True"
-
-if github_pages_build:
-    input_dir = "./src"
-    output_dir = "build"
-    configure_doxyfile(input_dir, output_dir)
-    subprocess.call("doxygen", shell=True)
-    breathe_projects["polyhedral-gravity-model"] = output_dir + "/xml"
+input_dir = "../src"
+output_dir = "build"
+configure_doxyfile(input_dir, output_dir)
+subprocess.call("doxygen", shell=True)
+breathe_projects["polyhedral-gravity-model"] = output_dir + "/xml"
 
 # -- Project information -----------------------------------------------------
 
