### 2024-03-25 accee67
improve script
