+++ .readthedocs.yaml
@@ -1,13 +1,0 @@
-# Read the Docs configuration file for Sphinx projects
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
-version: 2
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.12"
-sphinx:
-  configuration: docs/conf.py
-python:
- install:
-   - requirements: docs/requirements.txt
