### 2024-03-25 d9fad61
apply formulation suggestions of #32
