+++ paper/paper.md
@@ -45,7 +45,7 @@ The library supports many common formats, such as *.stl*, *.off*, *.ply*, *.mesh
 
 # Statement of Need
 
-The complex gravitational fields of irregular bodies, such as asteroids and comets, are often modeled using polyhedral gravity models since alternative approaches like mascon models or spherical harmonics struggle with these bodies' irregular geometry. The former struggles with convergence close to the surface [@vsprlak2021use], whereas the latter requires a computationally expensive amount of mascons (point masses of which the target body is composed) to model fine-granular surface geometry [@wittick2017mascon].
+The complex gravitational fields of irregular bodies, such as asteroids and comets, are often modeled using polyhedral gravity models since alternative approaches like mascon models or spherical harmonics struggle with these bodies' irregular geometry. The spherical harmonics approach struggles with convergence close to the surface [@vsprlak2021use], whereas mascon models require a computationally expensive amount of mascons (point masses of which the target body comprises) to model fine-granular surface geometry [@wittick2017mascon].
 
 In contrast, polyhedral gravity models provide an analytic solution for the computation of the gravitational potential, acceleration (and second derivative) given a mesh of the body [@tsoulis2012analytical;@tsoulis2021computational] with the only assumption of homogeneous density.
 The computation of the gravitational potential and acceleration is a computationally expensive task, especially for large meshes, which can however benefit from parallelization either over computed target points for which we seek potential and acceleration or over the mesh. Thus, a high-performance implementation of a polyhedral gravity model is desirable.
