### 2024-03-25 esa/polyhedral-gravity-model#32 @schuhmaj
Agreed, added by #31
