### 2024-03-25 19ea015
add environment.yml for Jupyter Notebook (#33)
