### 2024-03-25 be06023
environments for jupyter notebook
