### 2024-03-25 esa/polyhedral-gravity-model#34 @schuhmaj
Good point.

However, this check is intentionally not included in the `evaluate(..)` function due to its $O(n^2)$ runtime.
As background, we're using the Möller–Trumbore intersection algorithm. This algorithm computes whether a ray intersects a plane. The number of intersections must be even to check that a plane unit normally points outwards. This is done for every plane normal for every other plane.
The approach works for both convex and concave polyhedrons. The disadvantage is that it is quadratic, at least in the naiv implementation of the `utility` package.

An Improvement of the `utility` would be a future TODO. But even then, `O(n)`, as it is the runtime of the gravity model, is not reachable (to my knowledge).

--------
What I can do, however, is to improve the visibility of this check in the package.

Hence, I've stated the `mesh_checking` capabilities more clearly in the `README.md` to draw the attention of the user to this requirement. I've also slightly updated the documentation in this regard.

Resolved by #31

[1] https://stackoverflow.com/questions/45603469/how-to-calculate-the-normals-of-a-box
