### 2024-03-25 5d60f26
fix cmake generator selection
