### 2024-03-25 4d7d608
add badges to the docs as well
