+++ README.md
@@ -200,7 +200,7 @@ Binaries for the most common platforms are available on PyPI including
 Windows, Linux and macOS. For macOS and Linux, binaries for
 `x86_64` and `aarch64` are provided.
 In case `pip` uses the source distribution, please make sure that
-you have a C++17 capable compiler, CMake and ninja-build installed.
+you have a C++17 capable compiler and CMake installed.
 
 ### From source
 
