### 2024-03-26 714bda4
improve formatting given the expected input of doxygen
