+++ src/polyhedralGravity/calculation/MeshChecking.h
@@ -14,7 +14,7 @@ namespace polyhedralGravity::MeshChecking {
 
     /**
      * Checks if the vertices are in such a order that the unit normals of each plane point outwards the polyhedron
-     * @param polyhedron - the polyhedron consisting of vertices and triangular faces
+     * @param polyhedron the polyhedron consisting of vertices and triangular faces
      * @return true if all the unit normals are pointing outwards
      *
      * @note This has quadratic complexity @f$O(n^2)@f$! For bigger problem sizes, K-D trees/ Octrees could improve
@@ -26,7 +26,7 @@ namespace polyhedralGravity::MeshChecking {
     /**
      * Checks if no triangle is degenerated by checking the surface area being greater than zero.
      * E.g. two points are the same or all three are collinear.
-     * @param polyhedron - the polyhedron consisting of vertices and triangular faces
+     * @param polyhedron the polyhedron consisting of vertices and triangular faces
      * @return true if triangles are fine and none of them is degenerate
      */
     bool checkTrianglesNotDegenerated(const Polyhedron &polyhedron);
@@ -36,9 +36,9 @@ namespace polyhedralGravity::MeshChecking {
 
         /**
          * Calculates how often a vector starting at a specific origin intersects a polyhedron's mesh's triangles.
-         * @param rayOrigin - the origin of the ray
-         * @param rayVector - the vector describing the ray
-         * @param polyhedron - the polyhedron consisting of vertices and triangular faces
+         * @param rayOrigin the origin of the ray
+         * @param rayVector the vector describing the ray
+         * @param polyhedron the polyhedron consisting of vertices and triangular faces
          * @return true if the ray intersects the triangle
          */
         size_t
@@ -47,9 +47,9 @@ namespace polyhedralGravity::MeshChecking {
         /**
          * Calculates how often a vector starting at a specific origin intersects a triangular face.
          * Uses the Möller–Trumbore intersection algorithm.
-         * @param rayOrigin - the origin of the ray
-         * @param rayVector - the vector describing the ray
-         * @param triangle - a triangular face
+         * @param rayOrigin the origin of the ray
+         * @param rayVector the vector describing the ray
+         * @param triangle a triangular face
          * @return intersection point or null
          *
          * @related Adapted from https://en.wikipedia.org/wiki/Möller–Trumbore_intersection_algorithm
