+++ src/polyhedralGravity/util/UtilityThrust.h
@@ -7,8 +7,8 @@ namespace polyhedralGravity::util {
 
     /**
      * Returns a zip iterator for a pack of Iterators
-     * @tparam Iterator - should be a iterator
-     * @param args - can be any number of iterators
+     * @tparam Iterator should be a iterator
+     * @param args can be any number of iterators
      * @return a thrust::zip_iterator
      */
     template<typename ...Iterator>
@@ -18,8 +18,8 @@ namespace polyhedralGravity::util {
 
     /**
      * Returns a zip iterator for a pack of Containers
-     * @tparam Container - should be a container on which one can call std::begin()
-     * @param args - can be any number of containers
+     * @tparam Container should be a container on which one can call std::begin()
+     * @param args can be any number of containers
      * @return a thrust::zip_iterator for the beginning of all containers
      */
     template<typename ...Container>
@@ -29,8 +29,8 @@ namespace polyhedralGravity::util {
 
     /**
      * Returns a zip iterator for a pack of Containers
-     * @tparam Container - should be a container on which one can call std::end()
-     * @param args - can be any number of containers
+     * @tparam Container should be a container on which one can call std::end()
+     * @param args can be any number of containers
      * @return a thrust::zip_iterator for the end of all containers
      */
     template<typename ...Container>
@@ -40,8 +40,8 @@ namespace polyhedralGravity::util {
 
     /**
      * Returns a zip iterator pair for a pack of Containers
-     * @tparam Container - should be a container on which one can call std::begin() and std::end()
-     * @param args - can be any number of containers
+     * @tparam Container should be a container on which one can call std::begin() and std::end()
+     * @param args can be any number of containers
      * @return a pair of thrust::zip_iterator for the beginning and end of all containers
      */
     template<typename ...Container>

[…]
