+++ docs/quickstart/quick_start_cpp.rst
@@ -35,7 +35,7 @@ Further one must specify the name of the .csv output file.
         polyhedron: #polyhedron source-file(s)
           - "../example-config/data/tsoulis.node"   # .node contains the vertices
           - "../example-config/data/tsoulis.face"   # .face contains the triangular faces
-        density: 2670.0                             # constant density in [kg/m^3]
+        density: 2670.0                             # constant density, units must match with the mesh (see section below)
         points: # Location of the computation point(s) P
           - [ 0, 0, 0 ]                             # Here it is situated at the origin
         check_mesh: true                            # Fully optional, enables input checking (not given: false)
@@ -45,6 +45,21 @@ Further one must specify the name of the .csv output file.
 
 Have a look at :ref:`supported-polyhedron-source-files` to view the available
 options for polyhedral input.
+
+
+The calculation outputs the following parameters for every Computation Point *P*.
+The units of the respective output depend on the units of the input parameters (mesh and density)!
+Hence, if e.g. your mesh is in :math:`km`, the density must match. Further, output units will be different accordingly.
+
++------------------------------+----------------------------------------------------------------------------+-----------------------------------------------------------------+
+|         Name                 | If mesh :math:`[m]` and density :math:`[\frac{kg}{m^3}]`                   |                             Comment                             |
++==============================+============================================================================+=================================================================+
+|         V                    |  :math:`\frac{m^2}{s^2}` or :math:`\frac{J}{kg}`                           |           The potential or also called specific energy          |
++------------------------------+----------------------------------------------------------------------------+-----------------------------------------------------------------+
+|     Vx, Vy, Vz               |   :math:`\frac{m}{s^2}`                                                    |The gravitational acceleration in the three cartesian directions |
++------------------------------+----------------------------------------------------------------------------+-----------------------------------------------------------------+
+| Vxx, Vyy, Vzz, Vxy, Vxz, Vyz |   :math:`\frac{1}{s^2}`                                                    |The spatial rate of change of the gravitational acceleration     |
++------------------------------+----------------------------------------------------------------------------+-----------------------------------------------------------------+
 
 As Library
 ----------
@@ -128,12 +143,12 @@ about the vertices' ordering due to its quadratic complexity!
         }
 
 
-
-Evaluation with caching
-~~~~~~~~~~~~~~~~~~~~~~~
+GravityEvaluable (with caching)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Evaluating the gravity model for a given polyhedron
 defined from within source code for a specific point and density.
+
 
 .. code-block:: cpp
 

[…]
