### 2024-03-26 esa/polyhedral-gravity-model#34 @santisoler
I understand your point. I agree with making the `utility` module more discoverable through documentation.

What if the mesh check is added to the forward modelling function but also provide an optional argument to disable it. Since I would recommend every user to run the check before computing the gravity field, I would turn it on by default. In some particular situations users might not want to run the check , so they could do so by setting the argument to `False`.

Something like:
```python
potential, acceleration, tensor = polyhedral_gravity.evaluate(
    polyhedral_source=(cube_vertices, cube_faces),
    density=cube_density,
    computation_points=computation_point,
    parallel=True,
    disable_check=True,
)
```

What do you think?
