### 2024-03-26 esa/polyhedral-gravity-model#34 @schuhmaj
Agreed. I would even go a step further and enable the direct utilization with a "correct" mesh but reverse vertex ordering.
Further, I would add an interface to get a proper hint of where the vertex ordering is not aligned.

The algorithm requires the plane unit normals to be outwards pointing.
Hence, counterclockwise vertex ordering in a right-handed coordinate system.
I am not completely sure since this definition (clockwise/counterclockwise) also depends on the coordinate system.

Hence, I would just specify an enum "OUTWARDS" or "INWARDS" for the following feature.

# Changes

## Core Interface

- [ ] Add a parameter `polyhedral_orientation` to `evaluate(..)` --> `Enum Class NORMAL_ORIENTATION`
  - We implicitly assume always `OUTWARDS` ordering if not differently specified (Default Case)
  - The results are multiplied with `-1.0` if `INWARDS ` is specified
- [ ] Add a parameter `check_mesh` to `evaluate(..)` --> `optional<bool>`
  - The parameter is, by default, implicitly enabled and prints a warning that the mesh check is conducted requiring $O(n^2)$ (Default Case)
  - Explicitly setting the parameter to `True ` or `False ` silences the warning
  - In case the parameter is `True` a `runtime_error` is thrown if the mesh does not comply with the specified `polyhedral_orientation`
  - The method also conducts the $O(n)$ check for triangles without surface area
- [ ] Add the same behavior to `GravityEvaluable` in the constructor

## Utility Methods

- [ ] The `mesh_check` method is removed due to its integration into the core interface
- [ ] A method `check_vertex_order` is introduced, which returns
  - **Majority Ordering**
    - INWARDS --> All plane unit normals intersect the polyhedron odd times
    - OUTWARDS --> All plane unit normals intersect the polyhedron even times
  - **Indices of faces violating the order**
    - a list of face-indices violating the order/ might be empty if everything is "fine"
  The implementation determines which ordering is "more reasonable." If the majority of faces have inward-pointing normals, INWARD_POINTING is returned with "faulty" outward-pointing face indices, and vice versa.
  This way, I get the correct arguments for `polyhedral_orientation` and can abort if the list is non-empty (or exchange the vertex ordering at the returned faces)
