### 2024-03-26 esa/polyhedral-gravity-model#34 @schuhmaj
I will add these features in the next days, but in a new PR other than #31
