### 2024-03-26 esa/polyhedral-gravity-model#34 @santisoler
I heavily support opening a new PR for this.

Just a few minor thoughts:

> Explicitly setting the parameter to True or False silences the warning

I think setting it to `False` should prevent the check to being run, not only silencing the warning. This way users can choose to save computation time by disabling the checks (maybe that's what you meant, but I wanted to clarify it).

> The mesh_check method is removed due to its integration into the core interface

I wouldn't remove it. You already have it, and it's something useful to have. Specially while using the `evaluate` method. Maybe some user wants to manually make sure the mesh is correct before running the forward.

> Add the same behavior to GravityEvaluable in the constructor

That would be great! And I think you can only run the check once, no need to run it every time the user ask for the computation of the gravity fields. Maybe the check could be run on instantiation of the class. Or maybe check before the first time the gravity fields are computed, and store a private attribute that flags that the mesh has been checked and is valid.

I really like your approach here! Thanks a lot for considering my idea.
