### 2024-03-26 esa/polyhedral-gravity-model#28 @santisoler
> So, if I get this correctly, sorry, if I am asking so stupidly, but I am quite a lot confused with the sign:

There are not stupid questions 🙂. I think is super valuable that you are asking for clarification.

I would suggest to take the 1b path. I think it's easier to understand and document the definition of the potential as positive and that the acceleration vector is returned as:

$$ \mathbf{g} = \nabla V = \left( \frac{\partial V}{\partial x}, \frac{\partial V}{\partial y}, \frac{\partial V}{\partial z} \right)  $$

If the 1a path is chosen, then it would be wrong to call the `(-Vx, -Vy, -Vz)` vector the _acceleration_, because it's not any acceleration vector that could be observed in such scenario.

---

Minor comment regarding orientation of axis. I appreciate the point that @JoachimSchwabe makes regarding the gravity acceleration measurements in geophysics. Nonetheless, I think the main confusion that comes from (and this are historical reasons) the fact that in most geophysical literature, authors refer to $g_z$ as the **downward component** of the acceleration vector. By doing that mapping between $g_z$ and the downward direction, they were implicitly setting the $z$ axis pointing downwards. This imposed a problem when we need to generalize the forward models on systems where we don't have a natural _up_ or _down_ direction (asteroids for example), or even when our masses are above the observation point.

In conclusion, given the fact that `polyhedral-gravity-model` is mainly aimed to model bodies like asteroids, where there might not be a natural choice for the coordinate system, I suggest to follow a general formulation and keep definitions simpler.

From personal preference, I would use the geodetic convention of defining the potential $V$ as always positive for positive masses, and return the acceleration as the gradient of $V$ with respect to whatever x, y and z coordinates the user is defining (as long as the observation point and the mesh are defined in the same coordinate system). In summary, it would be the 1b path (as I stated before).
