### 2024-03-27 esa/polyhedral-gravity-model#28 @JoachimSchwabe
I pretty much agree with santosoler. Again, whether Vx etc. or -Vx etc. points towards the mass depends on the coordinate system and the location of mass and computation point. And as I had written before and santosoler pointed out again, the convention with the negative sign for the acceleration comes from the case of surface gravimetry where the acceleration is considered as the vertical component only and points downwards, whereas the height coordinate points upwards. This concept does not really make sense for general 3d coordinates, or also for the "horizontal" component of the gravity vector, if x and y are associated with east and north.
Thus, I would suggest to output positive V and Vx, Vy, Vz and tensor accordingly to prevent confusion and make a clear statement about it. It is the task of the user to e.g. take the absolute value of either component or the length of the gradient vector for the positive acceleration towards the mass, if that is needed, depending on the mass body and the coordinate system.
