### 2024-03-28 622fe7f
invert acceleration; upgrade version number; update description
