+++ docs/quickstart/examples_python.rst
@@ -1,5 +1,9 @@
-Quick Start Python
-==================
+.. _examples-python:
+
+Examples Python
+===============
+
+Details about mesh and input units can be found in :ref:`quick-start-io`.
 
 The use of the Python interface is pretty straight-forward since
 there is only one method: :code:`evaluate(..)` or the alternative
@@ -11,17 +15,7 @@ If you quickly want to try out something, use :code:`evaluate(..)`.
 The polyhedral source can either be a tuple of vertices and faces, or
 a list of polyhedral mesh files (see :ref:`supported-polyhedron-source-files`).
 
-.. note::
-
-    The plane unit normals of every face of the polyhedral mesh must point **outwards**
-    of the polyhedron!
-    You can check this property via :ref:`mesh-checking-cpp` in C++ or
-    via the :ref:`mesh-checking-python` submodule in Python.
-    If the vertex order of the faces is inverted, i.e. the plane unit normals point
-    inwards, then the sign of the output will be inverted.
-
-
-The method calls follow the same pattern as the C++ interface. Thus it is always:
+The method calls follow the same pattern as the C++ interface.
 
 .. code-block:: python
 
