+++ docs/background/approach.rst
@@ -1,0 +1,5 @@
+The Polyhedral Gravity Model
+============================
+
+
+This polyhedral gravity model

[…]
