+++ docs/quickstart/supported_input.rst
@@ -9,7 +9,7 @@ the polyhedral source. These include:
 ====================== ==================================================== ==================================================================================================================================================
 File Suffix            Name                                                 Comment
 ====================== ==================================================== ==================================================================================================================================================
-  `.node` and `.face`                     TetGen's files                     These two files need to be given as a pair to the input. [Documentation of TetGen's files](https://wias-berlin.de/software/tetgen/fformats.html)
+  `.node` and `.face`                     TetGen's files                     These two files need to be given as a pair to the input. `Documentation of TetGen's files <https://wias-berlin.de/software/tetgen/fformats.html>`__
         `.mesh`                         Medit's mesh files                   Single file containing every needed mesh information.
         `.ply`          The Polygon File format/ Stanfoard Triangle format   Single file containing every needed mesh information. Blender File Format.
         `.off`                          Object File Format                   Single file containing every needed mesh information.
