+++ README.md
@@ -49,7 +49,7 @@ and runtime measurements of the presented code.
 
 The implementation is based on the
 paper [Tsoulis, D., 2012. Analytical computation of the full gravity tensor of a homogeneous arbitrarily shaped polyhedral source using line integrals. Geophysics, 77(2), pp.F1-F11.](http://dx.doi.org/10.1190/geo2010-0334.1)
-and its corresponding [implementation in FORTRAN](https://software.seg.org/2012/0001/index.html).
+and its corresponding implementation in FORTRAN.
 
 Supplementary details can be found in the more recent
 paper [TSOULIS, Dimitrios; GAVRIILIDOU, Georgia. A computational review of the line integral analytical formulation of the polyhedral gravity signal. Geophysical Prospecting, 2021, 69. Jg., Nr. 8-9, S. 1745-1760.](https://doi.org/10.1111/1365-2478.13134)
