+++ docs/conf.py
@@ -70,7 +70,7 @@ html_theme = "sphinx_book_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ["_static"]
+html_static_path = ["_static", "figures"]
 
 # Breathe Configuration
 breathe_default_project = "polyhedral-gravity-model"
