+++ .github/workflows/docs.yml
@@ -22,5 +22,10 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y doxygen
+      - uses: actions/checkout@v3
+        with:
+          lfs: true
       - id: deployment
         uses: sphinx-notes/pages@v3
+        with:
+          checkout: false
