### 2024-03-28 esa/polyhedral-gravity-model#28 @schuhmaj
Thanks a lot to both of you for taking the time to detail your explanation a second time. Now, I get it 🙂

#31 fixes the sign issues by adding a -1 multiplication to the acceleration while also improving the documentation with regard to the convention being utilized.
