### 2024-04-08 2a60ca9
improve LaTeX syntax
Co-authored-by: Santiago Soler <santisoler@fastmail.com>
