### 2024-04-15 f6f7e33
move ray generation into the detail method
