### 2024-04-15 8389f55
Implement normals orientation check in MeshChecking
Modified the MeshChecking functionality to include an orientation check for both outward and inward facing normals. Also included a method to determine the majority vertex ordering of a polyhedron and which vertices violate this ordering. The new checks are implemented in the test suite and reflected in the main application.
