+++ src/polyhedralGravityPython/PolyhedralGravityPython.cpp
@@ -128,7 +128,7 @@ PYBIND11_MODULE(polyhedral_gravity, m) {
 
     utility.def("check_mesh", [](const std::vector<Array3> &vertices, const std::vector<IndexArray3> &faces) {
         Polyhedron poly{vertices, faces};
-        return MeshChecking::checkTrianglesNotDegenerated(poly) && MeshChecking::checkNormalsOutwardPointing(poly);
+        return MeshChecking::checkTrianglesNotDegenerated(poly) && MeshChecking::checkPlaneUnitNormalOrientation(poly);
     }, R"mydelimiter(
                 Checks if no triangles of the polyhedral mesh are degenerated by checking that their surface area
                 is greater zero.
@@ -145,7 +145,7 @@ PYBIND11_MODULE(polyhedral_gravity, m) {
     utility.def("check_mesh", [](const std::vector<std::string> &filenames) {
         TetgenAdapter tetgen{filenames};
         Polyhedron poly = tetgen.getPolyhedron();
-        return MeshChecking::checkTrianglesNotDegenerated(poly) && MeshChecking::checkNormalsOutwardPointing(poly);
+        return MeshChecking::checkTrianglesNotDegenerated(poly) && MeshChecking::checkPlaneUnitNormalOrientation(poly);
     }, R"mydelimiter(
                 Checks if no triangles of the polyhedral mesh are degenerated by checking that their surface area
                 is greater zero.
