### 2024-04-15 7547dd3
improve syntax of GitHub action
