+++ .github/workflows/build-and-test.yml
@@ -29,8 +29,12 @@ jobs:
           cd build
           cmake -DCMAKE_BUILD_TYPE=Release ..
           cmake --build . --config Release
-      - name: Test
+      - name: Test (Linux & macOS)
         run: cd build/test && ./polyhedralGravity_test
+        if: matrix.os != 'windows-latest'
+      - name: Test (Windows)
+        run: cd build/test && .\polyhedralGravity_test.exe
+        if: matrix.os == 'windows-latest'
 
   # Install the python interface & test that it is working using pytest
   build-and-test-python:
