### 2024-04-15 26c1dec
add new check parameter to the C++ implementation
