### 2024-04-15 c3fb332
dissolve model & calculation distinction - merge the components
