+++ src/polyhedralGravityPython/PolyhedralGravityPython.cpp
@@ -8,9 +8,9 @@
 
 #include "polyhedralGravity/model/Polyhedron.h"
 #include "polyhedralGravity/model/GravityModelData.h"
-#include "polyhedralGravity/calculation/GravityModel.h"
-#include "polyhedralGravity/calculation/GravityEvaluable.h"
-#include "polyhedralGravity/calculation/MeshChecking.h"
+#include "polyhedralGravity/model/GravityModel.h"
+#include "polyhedralGravity/model/GravityEvaluable.h"
+#include "polyhedralGravity/model/MeshChecking.h"
 #include "polyhedralGravity/input/TetgenAdapter.h"
 
 
