### 2024-04-16 4173c27
refactor the Polyhedron Class
Changes:
- add orientation from GravityEvaluable
- add density from GravityEvaluable
- add mesh checking from MeshChecking
- propagate changes inside C++ library/ executable

Rational:
All these methods/ functions were previously requiring to hand over a polyhedron. As instance methods, the argument is implicit (this) and we group mesh related utility into "the mesh container" and de-couple it from the gravity model.
