+++ src/polyhedralGravity/model/Polyhedron.cpp
@@ -1,0 +1,230 @@
+#include "Polyhedron.h"
+
+#include "polyhedralGravity/input/TetgenAdapter.h"
+
+
+namespace polyhedralGravity {
+
+    Polyhedron::Polyhedron(const std::vector<Array3> &vertices,
+            const std::vector<IndexArray3> &faces, double density, const NormalOrientation &orientation, const PolyhedronIntegrity &integrity)
+        : _vertices{vertices},
+          _faces{faces},
+          _density{density},
+          _orientation{orientation} {
+        //Checks that the node with index zero is actually used
+        if (_faces.end() == std::find_if(_faces.begin(), _faces.end(), [&](auto &face) {
+            return face[0] == 0 || face[1] == 0 || face[2] == 0;
+        })) {
+            throw std::invalid_argument("The node with index zero (0) was never used in any face! This is "
+                    "no valid polyhedron. Probable issue: Started numbering the vertices of "
+                    "the polyhedron at one (1).");
+        }
+        this->runIntegrityMeasures(integrity);
+    }
+
+    Polyhedron::Polyhedron(const PolyhedralSource &polyhedralSource, double density, const NormalOrientation &orientation, const PolyhedronIntegrity &integrity)
+        : Polyhedron{std::get<std::vector<Array3>>(polyhedralSource), std::get<std::vector<IndexArray3>>(polyhedralSource), density, orientation, integrity} {}
+
+    Polyhedron::Polyhedron(const PolyhedralFiles &polyhedralFiles, double density, const NormalOrientation &orientation, const PolyhedronIntegrity &integrity)
+        : Polyhedron{TetgenAdapter{polyhedralFiles}.getPolyhedron(), density, orientation, integrity} {}
+
+    const std::vector<Array3> &Polyhedron::getVertices() const {
+        return _vertices;
+    }
+
+    const Array3 &Polyhedron::getVertex(size_t index) const {
+        return _vertices[index];
+    }
+
+    size_t Polyhedron::countVertices() const {
+        return _vertices.size();
+    }
+
+    const std::vector<IndexArray3> &Polyhedron::getFaces() const {
+        return _faces;
+    }
+
+    const IndexArray3 &Polyhedron::getFace(size_t index) const {
+        return _faces[index];
+    }
+
+    Array3Triplet Polyhedron::getResolvedFace(size_t index) const {
+        return {_vertices[_faces[index][0]], _vertices[_faces[index][1]], _vertices[_faces[index][2]]};
+    }
+
+    size_t Polyhedron::countFaces() const {
+        return _faces.size();
+    }
+
+    double Polyhedron::getDensity() const {
+        return _density;
+    }
+
+    double Polyhedron::getOrientationFactor() const {
+        return _orientation == NormalOrientation::OUTWARDS ? 1.0 : -1.0;
+    }
+
+    std::string Polyhedron::toString() const {
+        std::stringstream sstream;
+        sstream << "<polyhedral_gravity.Polyhedron, density=" << _density << ", vertices= "
+           << countVertices() << ", faces= " << countFaces() << ", orientation= " << _orientation << ">";
+        return sstream.str();
+    }
+
+
+    void Polyhedron::runIntegrityMeasures(const PolyhedronIntegrity &integrity) {
+        using util::operator<<;
+        switch (integrity) {
+            case PolyhedronIntegrity::DISABLE:
+                return;
+            case PolyhedronIntegrity::AUTOMATIC:
+                SPDLOG_LOGGER_WARN(PolyhedralGravityLogger::DEFAULT_LOGGER.getLogger(),
+                                   "The mesh check is enabled and analyzes the polyhedron for degnerated faces & "
+                                   "that all plane unit normals point in the specified direction. This checks requires "
[…]
