### 2024-04-16 6522ca7
move polyhedron related enums into Polyhedron.h
