+++ src/polyhedralGravity/model/GravityModelData.h
@@ -49,76 +49,6 @@ namespace polyhedralGravity {
      */
     using GravityModelResult = std::tuple<double, Array3, Array6>;
 
-    /**
-     * The orientation of the plane unit normals of the polyhedron.
-     * We use this property as the concret definition of the vertices ordering depends on the
-     * utilized cooridnate system.
-     * However, the normal alignement is independent. Tsoulis et al. equations require the
-     * normals to point outwards of the polyhedron. If the opposite hold, the result is
-     * inverted.
-     */
-    enum class NormalOrientation: char {
-        /** Outwards pointing plane unit normals*/
-        OUTWARDS,
-        /** Inwards pointing plane unit normals */
-        INWWARDS
-    };
-
-
-    /**
-     * Overloaded insertion operator to output the string representation of a NormalOrientation enum value.
-     *
-     * @param os The output stream to write the string representation to.
-     * @param orientation The NormalOrientation enum value to output.
-     * @return The output stream after writing the string representation.
-     */
-    inline std::ostream& operator<<(std::ostream& os, const NormalOrientation& orientation) {
-        switch (orientation) {
-            case NormalOrientation::OUTWARDS:
-                os << "OUTWARDS";
-            break;
-            case NormalOrientation::INWWARDS:
-                os << "INWARDS";
-            break;
-            default:
-                os << "Unknown";
-            break;
-        }
-        return os;
-    }
-
-    /**
-     * The three mode the poylhedron class takes in
-     * the constructor in order to determine what initilaization checks to conduct.
-     * This enum is exclusivly utilized in the constrcutor of a {@link Polyhedron} and its private method
-     * {@link runIntegrityMeasures}
-     */
-    enum class PolyhedronIntegrity: char {
-        /**
-         * All activities regarding MeshChecking are disabled.
-         * @note No runtime overhead!
-         */
-        DISABLE,
-        /**
-         * Only verification of the NormalOrientation.
-         * A misalignment (e.g. specified OUTWARDS, but is not) leads to a runtime_error.
-         * @note Runtime Cost @f$O(n^2)$@f
-         */
-        VERIFY,
-        /**
-         * Like VERIFY, but also informs the user about the option in any case on the runtime costs.
-         * This is the implicit default option.
-         * @note Runtime Cost: @f$O(n^2)$@f and output to stdout in every case!
-         */
-        AUTOMATIC,
-        /**
-         * Verification and Autmatioc Healing of the NormalOrientation.
-         * A misalignemt does not lead to a runtime_error, but to an internal correction. Cost: @f$O(n^2)$@f
-         * @note Runtime Cost: @f$O(n^2)$@f and a modification of the mesh input!
-         */
-        HEAL,
-    };
-
 
     /**
      * Contains the 3D distances l1_pq and l2_pq between P and the endpoints of segment pq and
