### 2024-04-16 35a56d2
adapt test cases
