+++ test/model/GravityModelCubeTest.cpp
@@ -8,7 +8,6 @@
 #include <tuple>
 #include <fstream>
 #include <sstream>
-#include <string>
 #include "polyhedralGravity/model/GravityModel.h"
 #include "polyhedralGravity/model/Polyhedron.h"
 
@@ -20,37 +19,38 @@ class GravityModelCubeTest :
         public ::testing::TestWithParam<std::tuple<std::array<double, 3>, double, double, std::array<double, 3>>> {
 
 protected:
-
     /**
      * Small epsilon since we compare to ground truth
      */
     static constexpr double LOCAL_TEST_EPSILON = 1e-20;
-    const polyhedralGravity::Polyhedron _cube{{
-                                                      {-1.0, -1.0, -1.0},
-                                                      {1.0, -1.0, -1.0},
-                                                      {1.0, 1.0, -1.0},
-                                                      {-1.0, 1.0, -1.0},
-                                                      {-1.0, -1.0, 1.0},
-                                                      {1.0, -1.0, 1.0},
-                                                      {1.0, 1.0, 1.0},
-                                                      {-1.0, 1.0, 1.0}},
-                                              {
-                                                      {1,    3,    2},
-                                                      {0,   3,    1},
-                                                      {0,   1,   5},
-                                                      {0,    5,   4},
-                                                      {0,    7,    3},
-                                                      {0,   4,    7},
-                                                      {1,   2,   6},
-                                                      {1,    6,   5},
-                                                      {2, 3, 6},
-                                                      {3, 7, 6},
-                                                      {4, 5, 6},
-                                                      {4, 6, 7}}
+    polyhedralGravity::Polyhedron _cube{std::vector<polyhedralGravity::Array3>{
+                                                {-1.0, -1.0, -1.0},
+                                                {1.0, -1.0, -1.0},
+                                                {1.0, 1.0, -1.0},
+                                                {-1.0, 1.0, -1.0},
+                                                {-1.0, -1.0, 1.0},
+                                                {1.0, -1.0, 1.0},
+                                                {1.0, 1.0, 1.0},
+                                                {-1.0, 1.0, 1.0}},
+                                        std::vector<polyhedralGravity::IndexArray3>{
+                                                {1, 3, 2},
+                                                {0, 3, 1},
+                                                {0, 1, 5},
+                                                {0, 5, 4},
+                                                {0, 7, 3},
+                                                {0, 4, 7},
+                                                {1, 2, 6},
+                                                {1, 6, 5},
+                                                {2, 3, 6},
+                                                {3, 7, 6},
+                                                {4, 5, 6},
+                                                {4, 6, 7}},
+                                        1.0,
+                                        polyhedralGravity::NormalOrientation::OUTWARDS,
+                                        polyhedralGravity::PolyhedronIntegrity::DISABLE
     };
 
 public:
-
     [[nodiscard]] static std::vector<std::tuple<std::array<double, 3>, double, double, std::array<double, 3>>>
     readCubePoints(const std::string &filename) {
         std::vector<std::tuple<std::array<double, 3>, double, double, std::array<double, 3>>> result{};
@@ -76,8 +76,6 @@ public:
     }
 
 
-
-
 };
 
 TEST_P(GravityModelCubeTest, CubePoints) {
@@ -90,7 +88,8 @@ TEST_P(GravityModelCubeTest, CubePoints) {
[…]
