+++ test/model/PolyhedronTest.cpp
@@ -1,0 +1,312 @@
+#include "gtest/gtest.h"
+#include "gmock/gmock.h"
+
+#include <array>
+#include "polyhedralGravity/model/Polyhedron.h"
+
+
+/**
+ * Contains Tests if the mesh sanity checks of the Polyhedron work
+ * Parameterization not possible since the patterns what and where the error exactly is varies leading
+ * to different assertions being required in the single test case.
+ */
+class PolyhedronTest : public ::testing::Test {
+
+protected:
+    const std::vector<polyhedralGravity::Array3> _cubeVertices{
+            {-1.0, -1.0, -1.0},
+            {1.0, -1.0, -1.0},
+            {1.0, 1.0, -1.0},
+            {-1.0, 1.0, -1.0},
+            {-1.0, -1.0, 1.0},
+            {1.0, -1.0, 1.0},
+            {1.0, 1.0, 1.0},
+            {-1.0, 1.0, 1.0}
+    };
+
+    const std::vector<polyhedralGravity::IndexArray3> _facesOutwards{
+            {1, 3, 2},
+            {0, 3, 1},
+            {0, 1, 5},
+            {0, 5, 4},
+            {0, 7, 3},
+            {0, 4, 7},
+            {1, 2, 6},
+            {1, 6, 5},
+            {2, 3, 6},
+            {3, 7, 6},
+            {4, 5, 6},
+            {4, 6, 7}
+    };
+
+
+    const std::vector<polyhedralGravity::IndexArray3> _facesInwards{
+            {3, 1, 2},
+            {3, 0, 1},
+            {1, 0, 5},
+            {5, 0, 4},
+            {7, 0, 3},
+            {4, 0, 7},
+            {2, 1, 6},
+            {6, 1, 5},
+            {3, 2, 6},
+            {7, 3, 6},
+            {5, 4, 6},
+            {6, 4, 7}
+    };
+
+    const std::vector<polyhedralGravity::IndexArray3> _facesOutwardsMajority{
+            {3, 1, 2},
+            {0, 3, 1},
+            {0, 1, 5},
+            {0, 5, 4},
+            {7, 0, 3},
+            {0, 4, 7},
+            {1, 2, 6},
+            {1, 6, 5},
+            {2, 3, 6},
+            {3, 7, 6},
+            {4, 5, 6},
+            {4, 6, 7}
+    };
+
+    const std::vector<polyhedralGravity::IndexArray3> _facesInwardsMajority{
+            {3, 1, 2},
+            {3, 0, 1},
+            {1, 0, 5},
+            {5, 0, 4},
+            {7, 0, 3},
+            {4, 0, 7},
+            {2, 1, 6},
+            {6, 1, 5},
+            {3, 2, 6},
+            {3, 7, 6},
+            {4, 5, 6},
+            {4, 6, 7}
+    };
+
+    const std::vector<polyhedralGravity::IndexArray3> _degeneratedFaces{
+            {1, 3, 2},
+            {0, 3, 1},
+            {0, 1, 5},
+            {0, 5, 4},
+            {7, 7, 3},
+            {0, 4, 7},
+            {1, 2, 6},
+            {1, 6, 5},
+            {2, 3, 6},
+            {3, 7, 6},
+            {4, 5, 6},
+            {4, 6, 7}
+    };
+
+
+    const std::vector<polyhedralGravity::Array3> _prismVertices{
+            {-20.0, 0.0, 25.0},
+            {0.0, 0.0, 25.0},
[…]
