### 2024-04-16 e1a2276
new Python interface - part 1
