+++ src/polyhedralGravityPython/PolyhedralGravityPython.cpp
@@ -20,141 +20,263 @@ PYBIND11_MODULE(polyhedral_gravity, m) {
     m.doc() = "Computes the full gravity tensor for a given constant density polyhedron which consists of some "
             "vertices and triangular faces at a given computation points";
 
-    m.def("evaluate", [](const PolyhedralSource &polyhedralSource, double density,
-                         const std::variant<Array3, std::vector<Array3>> &computationPoints,
-                         bool parallel) -> std::variant<GravityModelResult, std::vector<GravityModelResult>> {
-              using namespace polyhedralGravity;
-              if (std::holds_alternative<Array3>(computationPoints)) {
-                  return std::visit([&density, &computationPoints, parallel](const auto &source) {
-                      using namespace polyhedralGravity;
-                      return GravityModel::evaluate(source, density, std::get<Array3>(computationPoints), parallel);
-                  }, polyhedralSource);
-              } else {
-                  return std::visit([&density, &computationPoints, parallel](const auto &source) {
-                      using namespace polyhedralGravity;
-                      return GravityModel::evaluate(source, density, std::get<std::vector<Array3>>(computationPoints),
-                                                    parallel);
-                  }, polyhedralSource);
-              }
-          }, R"mydelimiter(
-            Evaluates the polyhedral gravity model for a given constant density polyhedron at a given computation point.
+    py::enum_<NormalOrientation>(m, "NormalOrientation", R"mydelimiter(
+            The orientation of the plane unit normals of the polyhedron.
+            Tsoulis et al. equations require the normals to point outwards of the polyhedron.
+            If the opposite hold, the result is negated.
+            The implementation can handle both cases.
+        )mydelimiter")
+        .value("OUTWARDS", NormalOrientation::OUTWARDS, "Outwards pointing plane unit normals")
+        .value("INWARDS", NormalOrientation::INWWARDS, "Inwards pointing plane unit normals");
 
-            Args:
-                polyhedral_source:  The vertices & faces of the polyhedron as tuple or
-                                    the filenames of the files containing the vertices & faces as list of strings
-                density:            The constant density of the polyhedron in [kg/m^3]
-                computation_points: The computation points as tuple or list of points
-                parallel:           If true, the computation is done in parallel (default: true)
+    py::enum_<PolyhedronIntegrity>(m, "PolyhedronIntegrity", R"mydelimiter(
+            The pointing direction of the normals of a Polyhedron.
+            They can either point outwards or inwards the polyhedron.
+        )mydelimiter")
+        .value("DISABLE", PolyhedronIntegrity::DISABLE,
+            "All activities regarding MeshChecking are disabled. No runtime overhead!")
+        .value("VERIFY", PolyhedronIntegrity::VERIFY,
+            "Only verification of the NormalOrientation. "
+            "A misalignment (e.g. specified OUTWARDS, but is not) leads to a runtime_error. Runtime Cost O(n^2)")
+        .value("AUTOMATIC", PolyhedronIntegrity::AUTOMATIC,
+            "Like VERIFY, but also informs the user about the option in any case on the runtime costs. "
+            "This is the implicit default option. Runtime Cost: O(n^2) and output to stdout in every case!" )
+        .value("HEAL", PolyhedronIntegrity::HEAL,
+            "Verification and Autmatioc Healing of the NormalOrientation. "
+            "A misalignemt does not lead to a runtime_error, but to an internal correction of vertices ordering. Runtime Cost: @f$O(n^2)$@f");
 
-            Returns:
-                Either a tuple of potential, acceleration and second derivatives at the computation points or
-                if multiple computation points are given, the result is a list of tuples
[…]
