### 2024-04-16 3bb7843
new Python interface - part 2
