+++ src/polyhedralGravityPython/PolyhedralGravityPython.cpp
@@ -18,32 +18,32 @@ namespace py = pybind11;
 PYBIND11_MODULE(polyhedral_gravity, m) {
     using namespace polyhedralGravity;
     m.doc() = "Computes the full gravity tensor for a given constant density polyhedron which consists of some "
-            "vertices and triangular faces at a given computation points";
+            "vertices and triangular faces at a given computation points accoring to the geodetic convention";
 
     py::enum_<NormalOrientation>(m, "NormalOrientation", R"mydelimiter(
-            The orientation of the plane unit normals of the polyhedron.
-            Tsoulis et al. equations require the normals to point outwards of the polyhedron.
-            If the opposite hold, the result is negated.
-            The implementation can handle both cases.
+        The orientation of the plane unit normals of the polyhedron.
+        Tsoulis et al. equations require the normals to point outwards of the polyhedron.
+        If the opposite hold, the result is negated.
+        The implementation can handle both cases.
         )mydelimiter")
-        .value("OUTWARDS", NormalOrientation::OUTWARDS, "Outwards pointing plane unit normals")
-        .value("INWARDS", NormalOrientation::INWWARDS, "Inwards pointing plane unit normals");
+            .value("OUTWARDS", NormalOrientation::OUTWARDS, "Outwards pointing plane unit normals")
+            .value("INWARDS", NormalOrientation::INWWARDS, "Inwards pointing plane unit normals");
 
     py::enum_<PolyhedronIntegrity>(m, "PolyhedronIntegrity", R"mydelimiter(
-            The pointing direction of the normals of a Polyhedron.
-            They can either point outwards or inwards the polyhedron.
+        The pointing direction of the normals of a Polyhedron.
+        They can either point outwards or inwards the polyhedron.
         )mydelimiter")
-        .value("DISABLE", PolyhedronIntegrity::DISABLE,
-            "All activities regarding MeshChecking are disabled. No runtime overhead!")
-        .value("VERIFY", PolyhedronIntegrity::VERIFY,
-            "Only verification of the NormalOrientation. "
-            "A misalignment (e.g. specified OUTWARDS, but is not) leads to a runtime_error. Runtime Cost O(n^2)")
-        .value("AUTOMATIC", PolyhedronIntegrity::AUTOMATIC,
-            "Like VERIFY, but also informs the user about the option in any case on the runtime costs. "
-            "This is the implicit default option. Runtime Cost: O(n^2) and output to stdout in every case!" )
-        .value("HEAL", PolyhedronIntegrity::HEAL,
-            "Verification and Autmatioc Healing of the NormalOrientation. "
-            "A misalignemt does not lead to a runtime_error, but to an internal correction of vertices ordering. Runtime Cost: @f$O(n^2)$@f");
+            .value("DISABLE", PolyhedronIntegrity::DISABLE,
+                   "All activities regarding MeshChecking are disabled. No runtime overhead!")
+            .value("VERIFY", PolyhedronIntegrity::VERIFY,
+                   "Only verification of the NormalOrientation. "
+                   "A misalignment (e.g. specified OUTWARDS, but is not) leads to a runtime_error. Runtime Cost O(n^2)")
+            .value("AUTOMATIC", PolyhedronIntegrity::AUTOMATIC,
+                   "Like VERIFY, but also informs the user about the option in any case on the runtime costs. "
+                   "This is the implicit default option. Runtime Cost: O(n^2) and output to stdout in every case!")
+            .value("HEAL", PolyhedronIntegrity::HEAL,
+                   "Verification and Autmatioc Healing of the NormalOrientation. "
[…]
