### 2024-04-16 27f83ed
finalize the tests for the new Python interface
