### 2024-04-16 #6384 @schuhmaj
Hi @dfm

The "needed" requested changes are included in the paper and the implementation.
These are all integrated, yet the associated version `2.2` is not yet published since it changes behavior without a syntactical change. A potential existing user might run into trouble blindly upgrading.
Our current approach is to defer that a bit until I am finished - presumably finished this week - with https://github.com/esa/polyhedral-gravity-model/issues/36, which integrates an "optional" improvement that one will notice improving the first-time-user experience. I would ping you when it is merged & the changes are successfully released to `conda` and `PyPi`.

In any case, also at this place, thank you very much, @santisoler and @mikegrudic, for taking the time to review our submission 🙂 It got better by a lot!

## Summary

### Documentation related

- Improving `README.md` and `docs` by correcting examples and stating the relevant details more clearly
  - https://github.com/esa/polyhedral-gravity-model/issues/19
  - https://github.com/esa/polyhedral-gravity-model/issues/20
  - https://github.com/esa/polyhedral-gravity-model/issues/21
  - https://github.com/esa/polyhedral-gravity-model/issues/23
  - https://github.com/esa/polyhedral-gravity-model/issues/24
  - https://github.com/esa/polyhedral-gravity-model/issues/30
- Improving `CONTRIBUTING.md`
  - https://github.com/esa/polyhedral-gravity-model/issues/25
- Improving the `paper` formulations
  - https://github.com/esa/polyhedral-gravity-model/issues/32
- Documenting the `requirements` of the `jupyter example notebook`
  - https://github.com/esa/polyhedral-gravity-model/issues/33

### Implementation related

- Improving the build from source and relaxing the dependency on `ninja-build`
  - https://github.com/esa/polyhedral-gravity-model/issues/22
- Improving the test base (especially regarding the `density` parameter, which was previously always `1.0` being suboptimal in case of changes
  -  https://github.com/esa/polyhedral-gravity-model/issues/27
- **Major change** - Conmply correct with the geodetic convention of gravity. This required us to multiply our acceleration component with `-1.0`
  -  https://github.com/esa/polyhedral-gravity-model/issues/28

### (Optional) Implementation related - Work in Progress

What is still missing is an "optional" improvement
- https://github.com/esa/polyhedral-gravity-model/issues/34.

I am currently working on this issue in PR  https://github.com/esa/polyhedral-gravity-model/issues/36
