### 2024-04-16 #6384 @dfm
@schuhmaj — Thanks for this thorough summary! Let me know when https://github.com/esa/polyhedral-gravity-model/pull/36 is finished and we'll go through the final editorial steps.
