### 2024-04-17 125bdac
make it possible to build the docs from source locally without CMake
