### 2024-04-17 f7cb461
update documentation, make the Python docs autogenerated, add new pytests for the Polyhedron class
