+++ test/python/test_gravity_model.py
@@ -4,6 +4,7 @@ import numpy as np
 import pickle
 import pytest
 from pathlib import Path
+from functools import lru_cache
 
 CUBE_VERTICES = np.array([
     [-1, -1, -1],
@@ -31,16 +32,18 @@ CUBE_FACES = np.array([
     [4, 6, 7]
 ])
 
+CUBE_FACES_INVERTED = CUBE_FACES[:, [1, 0, 2]]
+
 DENSITY = 1.0
 
 CUBE_VERTICES_FILE = Path("test/resources/cube.node")
 CUBE_FACE_FILE = Path("test/resources/cube.face")
 
 
-@pytest.fixture(scope="session")
-def reference_solution() -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+@lru_cache(maxsize=10)
+def reference_solution(density: int) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
     """Reads the cube reference solution from a file and returns it as a tuple"""
-    reference_file_path = Path("test/resources/analytic_cube_solution_density1.txt")
+    reference_file_path = Path(f"test/resources/analytic_cube_solution_density{int(density)}.txt")
     reference = np.loadtxt(reference_file_path, skiprows=1)
     points = reference[:, :3]
     expected_potential = reference[:, 3].flatten()
@@ -49,20 +52,30 @@ def reference_solution() -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
 
 
 @pytest.mark.parametrize(
-    "polyhedral_source", [(CUBE_VERTICES, CUBE_FACES), [str(CUBE_VERTICES_FILE), str(CUBE_FACE_FILE)]],
-    ids=["with_arrays", "with_file_input"]
+    "polyhedral_source,normal_orientation,density", [
+        ((CUBE_VERTICES, CUBE_FACES), NormalOrientation.OUTWARDS, 1.0),
+        ((CUBE_VERTICES, CUBE_FACES_INVERTED), NormalOrientation.INWARDS, 1.0),
+        ([str(CUBE_VERTICES_FILE), str(CUBE_FACE_FILE)], NormalOrientation.OUTWARDS, 1.0),
+        ((CUBE_VERTICES, CUBE_FACES), NormalOrientation.OUTWARDS, 42.0),
+        ((CUBE_VERTICES, CUBE_FACES_INVERTED), NormalOrientation.INWARDS, 42.0),
+        ([str(CUBE_VERTICES_FILE), str(CUBE_FACE_FILE)], NormalOrientation.OUTWARDS, 42.0)
+    ],
+    ids=["with_arrays_outwards01", "with_arrays_inwards01", "with_file_input01",
+         "with_arrays_outwards42", "with_arrays_inwards42", "with_file_input42"]
 )
 def test_polyhedral_gravity(
         polyhedral_source: Union[Tuple[np.ndarray, np.ndarray], List[str]],
-        reference_solution: Tuple[np.ndarray, np.ndarray, np.ndarray]) -> None:
+        normal_orientation: NormalOrientation,
+        density: float
+) -> None:
     """Checks that the evaluate function the correct results and
     is callable with file/ array inputs.
     """
-    points, expected_potential, expected_acceleration = reference_solution
+    points, expected_potential, expected_acceleration = reference_solution(density)
     polyhedron = Polyhedron(
         polyhedral_source=polyhedral_source,
         density=DENSITY,
-        normal_orientation=NormalOrientation.OUTWARDS,
+        normal_orientation=normal_orientation,
         integrity_check=PolyhedronIntegrity.VERIFY,
     )
     sol = evaluate(
@@ -77,20 +90,30 @@ def test_polyhedral_gravity(
 
 
[…]
