### 2024-04-17 8bd1a4a
update jupyter notebook and everything else
