### 2024-04-18 e19d7ef
add new test case for Polyhedron Normal Orientation Detection
