### 2024-04-18 7886dc6
faster by 10x new checkPlaneUnitNormalOrientation routine using vector<bool>
