+++ docs/background/mesh_integrity.rst
@@ -70,7 +70,7 @@ for the construction of a polyhedron in the form of the enum :code:`PolyhedronIn
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
 | :code:`VERIFY`              | Throws an Exception if the :code:`NormalOrientation` is different or inconsistent than specified                                                                                                                        | NO                         | :math:`O(n^2)` |
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
-| :code:`HEAL`                | Modifies the specified :code:`NormalOrientation` and the faces array to a consistent polyhedron if is different or inconsistent                                                                                         | NO                         | :math:`O(n^2)` |
+| :code:`HEAL`                | Modifies the specified :code:`NormalOrientation` and the faces array to a consistent polyhedron if they are inconsistent                                                                                                | NO                         | :math:`O(n^2)` |
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
 | :code:`DISABLE`             | Disables all checks                                                                                                                                                                                                     | YES                        | None           |
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
