+++ .github/workflows/wheels.yml
@@ -39,7 +39,7 @@ jobs:
       # Building on macOS requires an installation of gcc since the default clang compiler
       # lacks certain features required for building the package
       - name: Build wheels (macOS)
-        uses: pypa/cibuildwheel@v2.15.0
+        uses: pypa/cibuildwheel@v2.17.0
         env:
           CIBW_BEFORE_BUILD: brew install ninja gcc@12
           CIBW_ENVIRONMENT: "CC=gcc-12 CXX=g++-12"
@@ -53,7 +53,7 @@ jobs:
       - uses: ilammy/msvc-dev-cmd@v1
         if: matrix.os == 'windows-latest'
       - name: Build wheels (Windows)
-        uses: pypa/cibuildwheel@v2.15.0
+        uses: pypa/cibuildwheel@v2.17.0
         env:
           CIBW_BEFORE_BUILD: choco install -y ninja cmake
           CIBW_ARCHS_WINDOWS: "auto64"
