### 2024-04-18 d333d31
version 3.0 - remove release candidate status
