### 2024-04-19 9519839
arm images in the pipeline [skip ci]
