+++ .github/workflows/wheels.yml
@@ -29,7 +29,7 @@ jobs:
         env:
           CIBW_BEFORE_BUILD: yum makecache && yum install -y gcc-c++ cmake && pip install ninja
           CIBW_BUILD: "*manylinux*"
-          CIBW_ARCHS_LINUX: "auto64"
+          CIBW_ARCHS_LINUX: "x86_64 aarch64"
           CIBW_TEST_COMMAND: 'python -c "import polyhedral_gravity"'
         with:
           package-dir: .
@@ -43,6 +43,7 @@ jobs:
         env:
           CIBW_BEFORE_BUILD: brew install ninja gcc@12
           CIBW_ENVIRONMENT: "CC=gcc-12 CXX=g++-12"
+          CIBW_ARCHS_MACOS: "x86_64 arm64"
           CIBW_TEST_COMMAND: 'python -c "import polyhedral_gravity"'
         with:
           package-dir: .
