### 2024-04-19 a0647d6
remove "install check" [skip ci]
