### 2024-04-19 927b203
include arm64 for macOS in PyPi pipeline
- aarch64 not working due to failure in x86 pipeline, dedicated aarch64 not available in GitHub runners at the moment
- arm64 working with Apple Clang compiler instead of gnu
