### 2024-04-21 de56197
update runtime measurements and add a plot
