+++ script/time.py
@@ -5,10 +5,17 @@ import matplotlib.pyplot as plt
 import mesh_utility
 import timeit
 import pickle
+from typing import Dict
 
 
-def main():
-    vertices, faces = mesh_utility.read_pk_file("/Users/schuhmaj/Programming/polyhedral-gravity-model/script/mesh/Eros.pk")
+def run_time_measurements(sample_size: int = 1000) -> Dict[str, float]:
+    """Returns the RuntimeMeasurements for the four configurations as mapping.
+
+    Returns:
+        a mapping from configuration name to runtime per point in microseconds
+    """
+    results = dict()
+    vertices, faces = mesh_utility.read_pk_file("./mesh/Eros.pk")
     vertices, faces = np.array(vertices), np.array(faces)
     polyhedron = Polyhedron(
         polyhedral_source=(vertices, faces),
@@ -17,48 +24,44 @@ def main():
     )
 
     # Generate 1000 random cartesian points
-    N = 1000
-    computation_points = np.random.uniform(-2, 2, (N, 3))
+    computation_points = np.random.uniform(-2, 2, (sample_size, 3))
 
-    ########################################################################################################################
-    # OLD
-    ########################################################################################################################
     start_time = timeit.default_timer()
-    for i in range(N):
+    for i in range(sample_size):
         evaluate(polyhedron, computation_points[i])
     end_time = timeit.default_timer()
 
-    delta = (end_time - start_time) / N * 1e6
+    delta = (end_time - start_time) / sample_size * 1e6
     # Print the time in milliseconds
     print("######## Old Single-Point ########")
-    print(f"--> {N} times 1 point with old interface")
+    print(f"--> {sample_size} times 1 point with old interface")
     print(f"--> Time taken: {delta:.3f} microseconds per point")
+    results[f"evaluate \n ${sample_size} \\times 1$ point"] = delta
 
     start_time = timeit.default_timer()
     evaluate(polyhedron, computation_points)
     end_time = timeit.default_timer()
 
-    delta = (end_time - start_time) / N * 1e6
+    delta = (end_time - start_time) / sample_size * 1e6
     # Print the time in milliseconds
     print("######## Old Multi-Point ########")
-    print("--> 1 time N points with old interface")
+    print(f"--> 1 time {sample_size} points with old interface")
     print(f"--> Time taken: {delta:.3f} microseconds per point")
+    results[f"evaluate \n $1 \\times {sample_size}$ points"] = delta
 
-    ########################################################################################################################
-    # NEW
-    ########################################################################################################################
     evaluable = GravityEvaluable(polyhedron)
 
     start_time = timeit.default_timer()
-    for i in range(N):
+    for i in range(sample_size):
         evaluable(computation_points[i])
     end_time = timeit.default_timer()
 
-    delta = (end_time - start_time) / N * 1e6
+    delta = (end_time - start_time) / sample_size * 1e6
     # Print the time in milliseconds
     print("######## GravityEvaluable (Single-Point) #########")
-    print(f"--> {N} times 1 point with GravityEvaluable")
+    print(f"--> {sample_size} times 1 point with GravityEvaluable")
     print(f"--> Time taken: {delta:.3f} microseconds per point")
+    results[f"GravityEvaluable \n ${sample_size} \\times 1$ point"] = delta
 
     evaluable = GravityEvaluable(polyhedron)
 
@@ -66,35 +69,39 @@ def main():
     evaluable(computation_points)
     end_time = timeit.default_timer()
 
-    delta = (end_time - start_time) / N * 1e6
+    delta = (end_time - start_time) / sample_size * 1e6
[…]
