+++ .github/workflows/build-and-test.yml
@@ -32,10 +32,6 @@ jobs:
       - name: Test (Linux & macOS)
         run: cd build/test && ./polyhedralGravity_test
         if: matrix.os != 'windows-latest'
-# Problems with resources (test file) location leads to test failure
-#      - name: Test (Windows)
-#        run: cd build/test/Release && .\polyhedralGravity_test.exe
-#        if: matrix.os == 'windows-latest'
 
   # Install the python interface & test that it is working using pytest
   build-and-test-python:
