### 2024-04-21 1aadac8
Apply suggestions from code review
Co-authored-by: Pablo Gómez <contact@pablo-gomez.net>
