### 2024-04-21 e9c32f4
Apply suggestions from code review
Co-authored-by: Pablo Gómez <contact@pablo-gomez.net>
