+++ docs/background/mesh_integrity.rst
@@ -66,7 +66,7 @@ for the construction of a polyhedron in the form of the enum :code:`PolyhedronIn
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
 | :code:`PolyhedronIntegrity` | Meaning                                                                                                                                                                                                                 | Invalid Polyhedron Possible| Overhead       |
 +=============================+=========================================================================================================================================================================================================================+============================+================+
-| :code:`AUTOMATIC` (default) | Throws an Exception if the :code:`NormalOrientation` is different or inconsistent than specified AND Prints a short-version of the above explaination to :code:`stdout` informing the user about :math:`O(n^2)` runtime | NO                         | :math:`O(n^2)` |
+| :code:`AUTOMATIC` (default) | Throws an Exception if the :code:`NormalOrientation` is different or inconsistent than specified AND Prints a short-version of the above explanation to :code:`stdout` informing the user about :math:`O(n^2)` runtime  | NO                         | :math:`O(n^2)` |
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
 | :code:`VERIFY`              | Throws an Exception if the :code:`NormalOrientation` is different or inconsistent than specified                                                                                                                        | NO                         | :math:`O(n^2)` |
 +-----------------------------+-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+----------------------------+----------------+
