### 2024-05-06 160a948
improve numerical stability of C++ intermediate test cases
- add method for calculating ULP distance
- Let the UtilityStructs use this method
- create custom matcher for comparing 2D containers
