### 2024-05-07 cbc508a
improves numerical test stability
This improves the stability of the tests and makes them independent of the utilized OS/ Compiler/ Architecture by:
- adding dedicated methods for FloatingPoint comparison using
  - UPL distance
  - Relative Epsilon
- adding custom GoogleTestMatchers using these new methods
- updating the DataStructures Equality methods
