+++ test/model/GravityModelBigTest.cpp
@@ -269,16 +269,15 @@ TEST_F(GravityModelBigTest, PlaneUnitNormals) {
 
     auto actualPlaneUnitNormals = polyhedralGravity::GravityModel::calculatePlaneUnitNormals(expectedGij);
 
-    ASSERT_THAT(actualPlaneUnitNormals, FloatContainter2D(expectedPlaneUnitNormals, 10));
+    ASSERT_THAT(actualPlaneUnitNormals, FloatContainter2D(expectedPlaneUnitNormals));
 }
 
 TEST_F(GravityModelBigTest, SegmentUnitNormals) {
     using namespace testing;
 
-    auto actualSegmentUnitNormals = polyhedralGravity::GravityModel::calculateSegmentUnitNormals(expectedGij,
-                                                                                                 expectedPlaneUnitNormals);
+    auto actualSegmentUnitNormals = polyhedralGravity::GravityModel::calculateSegmentUnitNormals(expectedGij, expectedPlaneUnitNormals);
 
-    ASSERT_THAT(actualSegmentUnitNormals, ContainerEq(expectedSegmentUnitNormals));
+    ASSERT_THAT(actualSegmentUnitNormals, FloatContainter3D(expectedSegmentUnitNormals));
 }
 
 TEST_F(GravityModelBigTest, PlaneNormalOrientations) {
@@ -304,7 +303,7 @@ TEST_F(GravityModelBigTest, PlaneDistances) {
 
     auto actualPlaneDistances = polyhedralGravity::GravityModel::calculatePlaneDistances(expectedHessianPlanes);
 
-    ASSERT_THAT(actualPlaneDistances, ContainerEq(expectedPlaneDistances));
+    ASSERT_THAT(actualPlaneDistances, FloatContainter1D(expectedPlaneDistances));
 }
 
 TEST_F(GravityModelBigTest, OrthogonalProjectionPointsOnPlane) {
@@ -363,7 +362,7 @@ TEST_F(GravityModelBigTest, SegmentDistances) {
             polyhedralGravity::GravityModel::calculateSegmentDistances(expectedOrthogonalProjectionPointsOnPlane,
                                                                        expectedOrthogonalProjectionPointsOnSegment);
 
-    ASSERT_THAT(actualSegmentDistances, ContainerEq(expectedSegmentDistances));
+    ASSERT_THAT(actualSegmentDistances, FloatContainter2D(expectedSegmentDistances));
 }
 
 TEST_F(GravityModelBigTest, DistancesPerSegmentEndpoint) {

[…]
