### 2024-05-07 f1f26d7
Update code documentation & Bump version number
Improved the code comments for better clarity and updated some of the examples in the docs. Also corrected minor typos in documentation and comments. The package version number is updated to 3.1.
