### 2024-05-08 f4ce167
Update src/polyhedralGravity/util/UtilityFloatArithmetic.h
Co-authored-by: Pablo Gómez <contact@pablo-gomez.net>
