### 2024-05-08 5b7d797
use infinity
