+++ cmake/pybind11.cmake
@@ -2,7 +2,7 @@ include(FetchContent)
 
 message(STATUS "Setting up pybind11")
 
-find_package(pybind11 2.10.4 QUIET)
+find_package(pybind11 2.12.0 QUIET)
 
 if (${pybind11_FOUND})
 
@@ -12,11 +12,10 @@ else()
 
     message(STATUS "Using pybind11 from git repository")
 
-    #Fetches the version 2.10.4 from the official github of pybind11
     FetchContent_Declare(pybind11
             GIT_REPOSITORY https://github.com/pybind/pybind11
-            GIT_TAG v2.10.4
-            )
+            GIT_TAG v2.12.0
+    )
 
     FetchContent_MakeAvailable(pybind11)
 
