### 2024-05-15 75a0732
decrease spdlog version to 1.13.0
spdlog 1.14.1 an 1.14.0 are not yet
available on conda-forge. This might
cause issues for the conda-deployment
otherwise
