### 2024-05-20 #6384 @schuhmaj
@dfm The final version for the JOSS release is `3.2.0`
The source code is published on Zenodo:

Schuhmacher, J., Blazquez, E., Gratl, F., Izzo, D., & Gómez, P. (2024). Efficient Polyhedral Gravity Modeling in Modern C++ and Python (3.2.0). Zenodo. https://doi.org/10.5281/zenodo.11221939

Ready for the next steps 🙂
