### 2024-06-01 #6384 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
